refactor(constants): reorganize split into themed per language and per concern submodules
