+++ codebase_rag/constants/__init__.py
@@ -1,3073 +1,27 @@
-# (H) Constants package: themed submodules re-exported for backwards compatibility.
+# (H) Application constants, re-exported from themed submodules.
 
-from enum import StrEnum
-from typing import NamedTuple
-
-from .ast_nodes import (
-    BINDINGS_DIR,
-    BUILD_EXT_CMD,
-    CAPTURE_ARROW_FUNCTION,
-    CAPTURE_CALL,
-    CAPTURE_CHILD_CLASS,
-    CAPTURE_CLASS,
-    CAPTURE_CONSTRUCTOR_NAME,
-    CAPTURE_EXPORT_FUNCTION,
-    CAPTURE_EXPORT_NAME,
-    CAPTURE_EXPORTS_OBJ,
-    CAPTURE_EXPORTS_PROP,
-    CAPTURE_FUNC,
-    CAPTURE_FUNCTION,
-    CAPTURE_FUNCTION_EXPR,
-    CAPTURE_IMPORT,
-    CAPTURE_IMPORT_FROM,
-    CAPTURE_MEMBER_EXPR,
-    CAPTURE_METHOD_FUNCTION,
-    CAPTURE_METHOD_NAME,
-    CAPTURE_MODULE_OBJ,
-    CAPTURE_PARENT_CLASS,
-    CAPTURE_PROTOTYPE_KEYWORD,
-    CAPTURE_VARIABLE_DECLARATOR,
-    CLI_STATS_TOTAL_NODES,
-    CPP_IMPORT_NODES,
-    FIELD_ALIAS,
-    FIELD_API_KEY,
-    FIELD_ARGUMENTS,
-    FIELD_BODY,
-    FIELD_BOUNDS,
-    FIELD_CONSTRUCTOR,
-    FIELD_DECLARATOR,
-    FIELD_DEFINITION,
-    FIELD_ENDPOINT,
-    FIELD_FIELD,
-    FIELD_FUNCTION,
-    FIELD_INTERFACES,
-    FIELD_KEY,
-    FIELD_LEFT,
-    FIELD_MODEL_ID,
-    FIELD_MODULE_NAME,
-    FIELD_NAME,
-    FIELD_OBJECT,
-    FIELD_OPERAND,
-    FIELD_OPERATOR,
-    FIELD_PARAMETERS,
-    FIELD_PROPERTY,
-    FIELD_PROVIDER,
-    FIELD_RECEIVER,
-    FIELD_RESULT,
-    FIELD_RETURN_TYPE,
-    FIELD_RIGHT,
-    FIELD_SCOPE,
-    FIELD_SUPERCLASS,
-    FIELD_SUPERCLASSES,
-    FIELD_TRAIT,
-    FIELD_TYPE,
-    FIELD_VALUE,
-    GRAMMARS_DIR,
-    INDEX_INDEX,
-    INDEX_INIT,
-    INDEX_MOD,
-    INPLACE_FLAG,
-    JS_LOCALS_PATTERN,
-    JS_TS_CLASS_NODES,
-    JS_TS_FUNCTION_NODES,
-    JS_TS_IMPORT_NODES,
-    KEY_NODES,
-    KEY_TOTAL_NODES,
-    LANG_ATTR_PHP,
-    LANG_ATTR_PREFIX,
-    LANG_ATTR_TSX,
-    LANG_ATTR_TYPESCRIPT,
-    LANG_DEFAULT_CALL_NODES,
-    LANG_DEFAULT_CLASS_NODES,
-    LANG_DEFAULT_FUNCTION_NODES,
-    LANG_DEFAULT_MODULE_NODES,
-    LANG_MSG_AVAILABLE_NODES,
-    METHOD_FIND_WITH_PREFIX,
-    METHOD_ITEMS,
-    NAME_FIELDS,
-    QUERY_CALLS,
-    QUERY_CAPTURE_CLASS,
-    QUERY_CAPTURE_FUNCTION,
-    QUERY_CLASSES,
-    QUERY_CONFIG,
-    QUERY_FUNCTIONS,
-    QUERY_IMPORTS,
-    QUERY_KEY_CLASSES,
-    QUERY_KEY_FUNCTIONS,
-    QUERY_LANGUAGE,
-    QUERY_LOCALS,
-    QUERY_NOT_AVAILABLE,
-    QUERY_RESULTS_PANEL_TITLE,
-    QUERY_SUMMARY_DB_ERROR,
-    QUERY_SUMMARY_SUCCESS,
-    QUERY_SUMMARY_TIMEOUT,
-    QUERY_SUMMARY_TRANSLATION_FAILED,
-    QUERY_SUMMARY_TRUNCATED,
-    SETUP_PY,
-    TREE_SITTER_MODULE_PREFIX,
-    TREE_SITTER_PREFIX,
-    TS_LOCALS_PATTERN,
-    TS_RS_TRAIT_BOUNDS,
-    TreeSitterModule,
-)
-from .cli import (
-    _SYSTEM_DIRS_PATTERN as _SYSTEM_DIRS_PATTERN,
-)
-from .cli import (
-    CLI_DEADCODE_COL_KIND,
-    CLI_DEADCODE_COL_LINES,
-    CLI_DEADCODE_COL_QUALIFIED_NAME,
-    CLI_DEADCODE_CONNECTING,
-    CLI_DEADCODE_LINE_RANGE,
-    CLI_DEADCODE_NONE,
-    CLI_DEADCODE_SUMMARY,
-    CLI_DEADCODE_TABLE_TITLE,
-    CLI_DEADCODE_WRITTEN,
-    CLI_ERR_CONFIG,
-    CLI_ERR_DEADCODE_AMBIGUOUS_PROJECT,
-    CLI_ERR_DEADCODE_FAILED,
-    CLI_ERR_DEADCODE_NO_PROJECTS,
-    CLI_ERR_DELETE_PROJECT_FAILED,
-    CLI_ERR_EXPORT_FAILED,
-    CLI_ERR_INDEXING,
-    CLI_ERR_JSON_REQUIRES_ASK_AGENT,
-    CLI_ERR_LOAD_GRAPH,
-    CLI_ERR_MCP_SERVER,
-    CLI_ERR_ONLY_JSON,
-    CLI_ERR_OUTPUT_REQUIRES_UPDATE,
-    CLI_ERR_PATH_NOT_DIR,
-    CLI_ERR_PATH_NOT_EXISTS,
-    CLI_ERR_PROJECT_NAME_REQUIRED,
-    CLI_ERR_PROJECT_NOT_FOUND,
-    CLI_ERR_STARTUP,
-    CLI_ERR_STATS_FAILED,
-    CLI_MSG_APP_TERMINATED,
-    CLI_MSG_AUTO_EXCLUDE,
-    CLI_MSG_CLEAN_DONE,
-    CLI_MSG_CLEANING_DB,
-    CLI_MSG_CLEANING_HASH_CACHE,
-    CLI_MSG_CONNECTING_MEMGRAPH,
-    CLI_MSG_CONNECTING_STATS,
-    CLI_MSG_DELETING_PROJECT,
-    CLI_MSG_EXPORTING_DATA,
-    CLI_MSG_EXPORTING_TO,
-    CLI_MSG_GRAPH_SUMMARY,
-    CLI_MSG_GRAPH_UPDATED,
-    CLI_MSG_HINT_TARGET_REPO,
-    CLI_MSG_INDEXING_AT,
-    CLI_MSG_INDEXING_DONE,
-    CLI_MSG_MCP_TERMINATED,
-    CLI_MSG_OPTIMIZATION_TERMINATED,
-    CLI_MSG_OUTPUT_TO,
-    CLI_MSG_PROJECT_DELETED,
-    CLI_MSG_SYNC_DONE,
-    CLI_MSG_SYNC_SKIPPED,
-    CLI_MSG_SYNCING_GRAPH,
-    CLI_MSG_UPDATING_GRAPH,
-    CLI_MSG_VERSION,
-    CLI_MSG_WORKSPACE_EMPTY,
-    CLI_MSG_WORKSPACE_SYNC_REPO,
-    CLI_MSG_WORKSPACE_SYNCING,
-    CLI_STATS_COL_COUNT,
-    CLI_STATS_COL_NODE_TYPE,
-    CLI_STATS_COL_REL_TYPE,
-    CLI_STATS_NODE_TITLE,
-    CLI_STATS_REL_TITLE,
-    CLI_STATS_TOTAL_RELS,
-    CLI_STATS_UNKNOWN,
-    CLI_WARN_NOT_GIT_REPO,
-    CONFIRM_DISABLED,
-    CONFIRM_ENABLED,
-    DEFAULT_API_KEY,
-    DEFAULT_DEBOUNCE_SECONDS,
-    DEFAULT_MAX_WAIT_SECONDS,
-    DEFAULT_MODEL,
-    DEFAULT_MODEL_ROLE,
-    DEFAULT_REGION,
-    DEFAULT_TABLE_TITLE,
-    DIFF_CONTINUATION_PREFIXES,
-    DIFF_FALLBACK_PATH,
-    DIFF_GIT_HEADER,
-    DIFF_LABEL_AFTER,
-    DIFF_LABEL_BEFORE,
-    ENV_ANTHROPIC_API_KEY,
-    ENV_AZURE_API_KEY,
-    ENV_AZURE_API_VERSION,
-    ENV_AZURE_ENDPOINT,
-    ENV_GOOGLE_API_KEY,
-    ENV_OPENAI_API_KEY,
-    EXIT_COMMANDS,
-    GREP_SUGGESTION,
-    HELP_ARG,
-    HELP_COMMAND,
-    HORIZONTAL_SEPARATOR,
-    INTERACTIVE_BFS_MAX_DEPTH,
-    INTERACTIVE_COL_NESTED,
-    INTERACTIVE_COL_NUM,
-    INTERACTIVE_COL_PATH,
-    INTERACTIVE_COL_PATTERN,
-    INTERACTIVE_DEFAULT_GROUP,
-    INTERACTIVE_EXPAND_SUFFIX,
-    INTERACTIVE_INSTRUCTIONS_GROUPED,
-    INTERACTIVE_INSTRUCTIONS_NESTED,
-    INTERACTIVE_KEEP_ALL,
-    INTERACTIVE_KEEP_NONE,
-    INTERACTIVE_NESTED_PLURAL,
-    INTERACTIVE_NESTED_SINGULAR,
-    INTERACTIVE_PROMPT_KEEP,
-    INTERACTIVE_STATUS_CGRIGNORE,
-    INTERACTIVE_STATUS_CLI,
-    INTERACTIVE_STATUS_DETECTED,
-    INTERACTIVE_STYLE_DIM,
-    INTERACTIVE_TITLE_GROUPED,
-    INTERACTIVE_TITLE_NESTED,
-    LOG_FORMAT,
-    LOG_LEVEL_ERROR,
-    LOG_LEVEL_INFO,
-    MODEL_COMMAND_PREFIX,
-    MSG_CHAT_INSTRUCTIONS,
-    MSG_CONNECTED_MEMGRAPH,
-    MSG_DOC_NO_CANDIDATES,
-    MSG_DOC_NO_CONTENT,
-    MSG_SEMANTIC_NO_RESULTS,
-    MSG_SEMANTIC_RESULT_FOOTER,
-    MSG_SEMANTIC_RESULT_HEADER,
-    MSG_SEMANTIC_SOURCE_FORMAT,
-    MSG_SEMANTIC_SOURCE_UNAVAILABLE,
-    MSG_SURGICAL_FAILED,
-    MSG_SURGICAL_SUCCESS,
-    MSG_SYNCING_KNOWLEDGE_GRAPH,
-    MSG_SYNCING_WORKSPACE,
-    MSG_THINKING_CANCELLED,
-    MSG_TIMEOUT_FORMAT,
-    MSG_TOOL_CALL_CANCELLED,
-    MULTILINE_INPUT_HINT,
-    NO_ANSWERS,
-    OPTIMIZATION_TABLE_TITLE,
-    PERMISSION_MODE_NORMAL_LABEL,
-    PERMISSION_MODE_TOGGLED,
-    PERMISSION_MODE_YOLO_LABEL,
-    PROMPT_ASK_QUESTION,
-    PROMPT_YOUR_RESPONSE,
-    REALTIME_LOGGER_FORMAT,
-    SESSION_CONTEXT_END,
-    SESSION_CONTEXT_START,
-    SESSION_LOG_EXT,
-    SESSION_LOG_HEADER,
-    SESSION_LOG_PREFIX,
-    SESSION_PREFIX_ASSISTANT,
-    SESSION_PREFIX_USER,
-    SHELL_CMD_GIT,
-    SHELL_CMD_GREP,
-    SHELL_CMD_RM,
-    SHELL_CMD_WHERE,
-    SHELL_CMD_WHICH,
-    SHELL_DANGEROUS_COMMANDS,
-    SHELL_DANGEROUS_PATTERNS_PIPELINE,
-    SHELL_DANGEROUS_PATTERNS_SEGMENT,
-    SHELL_PIPE_OPERATORS,
-    SHELL_REDIRECT_OPERATORS,
-    SHELL_RETURN_CODE_ERROR,
-    SHELL_RM_DANGEROUS_FLAGS,
-    SHELL_RM_FORCE_FLAG,
-    SHELL_RM_RF_FLAG,
-    SHELL_SUBSHELL_PATTERNS,
-    SHELL_SYSTEM_DIRECTORIES,
-    STATUS_BAR_BRANCH_CLEAN_HTML,
-    STATUS_BAR_BRANCH_CLEAN_PLAIN,
-    STATUS_BAR_BRANCH_DIRTY_HTML,
-    STATUS_BAR_BRANCH_DIRTY_PLAIN,
-    STATUS_BAR_BRANCH_RICH_TEXT,
-    STATUS_BAR_CLEAN_STYLE,
-    STATUS_BAR_CONFIG_COLOR,
-    STATUS_BAR_CONFIG_LABEL_C,
-    STATUS_BAR_CONFIG_LABEL_COLOR,
-    STATUS_BAR_CONFIG_LABEL_EDIT,
-    STATUS_BAR_CONFIG_LABEL_INSTRUCTIONS,
-    STATUS_BAR_CONFIG_LABEL_O,
-    STATUS_BAR_CONFIG_LABEL_REPO,
-    STATUS_BAR_CONFIG_SEPARATOR,
-    STATUS_BAR_DIRTY_MARKER,
-    STATUS_BAR_DIRTY_STYLE,
-    STATUS_BAR_EDIT_OFF,
-    STATUS_BAR_EDIT_ON,
-    STATUS_BAR_SEPARATOR_CHAR,
-    STATUS_BAR_SEPARATOR_COLOR,
-    STATUS_BAR_SPINNER,
-    STATUS_BAR_TOKEN_HTML,
-    TABLE_COL_CONFIGURATION,
-    TABLE_COL_VALUE,
-    TABLE_ROW_CYPHER_MODEL,
-    TABLE_ROW_EDIT_CONFIRMATION,
-    TABLE_ROW_OLLAMA_CYPHER,
-    TABLE_ROW_OLLAMA_ENDPOINT,
-    TABLE_ROW_OLLAMA_ORCHESTRATOR,
-    TABLE_ROW_ORCHESTRATOR_MODEL,
-    TABLE_ROW_TARGET_LANGUAGE,
-    TABLE_ROW_TARGET_REPOSITORY,
-    TMP_DIR,
-    TOKEN_COLOR_CRITICAL,
-    TOKEN_COLOR_OK,
-    TOKEN_COLOR_WARNING,
-    TOKEN_THRESHOLD_CRITICAL,
-    TOKEN_THRESHOLD_WARNING,
-    UI_DIFF_FILE_HEADER,
-    UI_ERR_EXPORT_FAILED,
-    UI_ERR_UNEXPECTED,
-    UI_FEEDBACK_PROMPT,
-    UI_GRAPH_EXPORT_STATS,
-    UI_GRAPH_EXPORT_SUCCESS,
-    UI_HELP_COMMANDS,
-    UI_INPUT_PROMPT_HTML,
-    UI_MODEL_CURRENT,
-    UI_MODEL_SWITCH_ERROR,
-    UI_MODEL_SWITCHED,
-    UI_MODEL_USAGE,
-    UI_NEW_FILE_HEADER,
-    UI_OPTIMIZATION_INIT,
-    UI_OPTIMIZATION_PANEL,
-    UI_OPTIMIZATION_START,
-    UI_REFERENCE_DOC_INFO,
-    UI_SHELL_COMMAND_HEADER,
-    UI_TOOL_APPROVAL,
-    UI_TOOL_ARGS_FORMAT,
-    WATCHER_SLEEP_INTERVAL,
-    YES_ANSWER,
-    YES_ANSWERS,
-    Color,
-    DiffMarker,
-    FileAction,
-    GoogleProviderType,
-    ModelRole,
-    PermissionMode,
-    Provider,
-    StyleModifier,
-)
-from .core import (
-    ABSTRACT_DECORATORS,
-    ALLOWED_COMMENT_MARKERS,
-    ANTHROPIC_API_VERSION,
-    ANTHROPIC_COUNT_TIMEOUT_S,
-    ANTHROPIC_COUNT_TOKENS_URL,
-    ANTHROPIC_HEADER_API_KEY,
-    ANTHROPIC_HEADER_VERSION,
-    ARG_COMMAND,
-    ARG_CONTENT,
-    ARG_FILE_PATH,
-    ARG_REPLACEMENT_CODE,
-    ARG_TARGET_CODE,
-    ATTR_TYPE_INFERENCE_IN_PROGRESS,
-    AUDIT_DETAIL_DANGLING,
-    AUDIT_DETAIL_MISSING_REQUIRED,
-    AUDIT_DETAIL_MISSING_REQUIRED_LIVE,
-    AUDIT_DETAIL_ORPHAN,
-    AUDIT_DETAIL_ORPHAN_COUNT,
-    AUDIT_DETAIL_UNDOCUMENTED_LABEL,
-    AUDIT_DETAIL_UNDOCUMENTED_PROPERTY,
-    AUDIT_DETAIL_UNDOCUMENTED_PROPERTY_LIVE,
-    AUDIT_DETAIL_UNDOCUMENTED_RELATIONSHIP,
-    BINARY_FILE_PERMISSION,
-    BINARY_NAME_TEMPLATE,
-    BUILTIN_PREFIX,
-    BYTES_PER_MB,
-    BYTES_PER_MB_FLOAT,
-    C_NAME_NODE_TYPES,
-    CALL_NODES_BASIC,
-    CALL_NODES_METHOD,
-    CALL_NODES_OPERATOR,
-    CALL_NODES_SPECIAL,
-    CHAR_ANGLE_CLOSE,
-    CHAR_ANGLE_OPEN,
-    CHAR_COLON,
-    CHAR_COMMA,
-    CHAR_HYPHEN,
-    CHAR_PAREN_CLOSE,
-    CHAR_PAREN_OPEN,
-    CHAR_SEMICOLON,
-    CHAR_SPACE,
-    CHAR_UNDERSCORE,
-    CLASS_NODES_BASIC,
-    CLASS_NODES_ENUM,
-    CLASS_NODES_INTERFACE,
-    CLASS_NODES_STRUCT,
-    CLASS_NODES_TYPE_ALIAS,
-    CLASS_NODES_UNION,
-    COMMENT_CHAR,
-    CONTENT_TYPE_JSON,
-    CPP_CLASS_TYPES,
-    CPP_COMPOUND_TYPES,
-    CPP_DECLARATOR_SUFFIX,
-    CPP_DESTRUCTOR_PREFIX,
-    CPP_EXPORT_CLASS_PREFIX,
-    CPP_EXPORT_MODULE_PREFIX,
-    CPP_EXPORT_PREFIXES,
-    CPP_EXPORT_STRUCT_PREFIX,
-    CPP_EXPORT_TEMPLATE_PREFIX,
-    CPP_EXPORT_UNION_PREFIX,
-    CPP_EXPORTED_CLASS_KEYWORDS,
-    CPP_FALLBACK_DESTRUCTOR,
-    CPP_FALLBACK_OPERATOR,
-    CPP_IMPL_SUFFIX,
-    CPP_KEYWORD_CLASS,
-    CPP_KEYWORD_STRUCT,
-    CPP_MODULE_PATH_MARKERS,
-    CPP_MODULE_PREFIX,
-    CPP_MODULE_PRIVATE_PREFIX,
-    CPP_MODULE_TYPE_IMPLEMENTATION,
-    CPP_MODULE_TYPE_INTERFACE,
-    CPP_NAME_NODE_TYPES,
-    CPP_NESTED_SCOPE_NODE_TYPES,
-    CPP_OPERATOR_SYMBOL_MAP,
-    CPP_OPERATOR_TEXT_PREFIX,
-    CPP_PREFIX_HAS,
-    CPP_PREFIX_IS,
-    CPP_STD_NAMESPACE,
-    CPP_STD_PREFIX,
-    CPP_STDLIB_ENTITIES,
-    CPP_TYPE_PARAMETER_DECL_TYPES,
-    CPP_TYPE_SPECIFIER_NODE_TYPES,
-    CSPROJ_SUFFIX,
-    CYPHER_ALL_MODULE_QNS,
-    DECORATOR_AT,
-    DEFAULT_CONTEXT_WINDOW,
-    DEFAULT_NAME,
-    DELIMITER_TOKENS,
-    DEP_ATTR_INCLUDE,
-    DEP_ATTR_VERSION,
-    DEP_EXCLUDE_PHP,
-    DEP_EXCLUDE_PYTHON,
-    DEP_FILE_CARGO,
-    DEP_FILE_COMPOSER,
-    DEP_FILE_GEMFILE,
-    DEP_FILE_GOMOD,
-    DEP_FILE_PACKAGE_JSON,
-    DEP_FILE_PYPROJECT,
-    DEP_FILE_REQUIREMENTS,
-    DEP_KEY_DEPENDENCIES,
-    DEP_KEY_DEV_DEPENDENCIES,
-    DEP_KEY_DEV_DEPS_JSON,
-    DEP_KEY_GROUP,
-    DEP_KEY_OPTIONAL_DEPS,
-    DEP_KEY_PEER_DEPS,
-    DEP_KEY_POETRY,
-    DEP_KEY_PROJECT,
-    DEP_KEY_REQUIRE,
-    DEP_KEY_REQUIRE_DEV,
-    DEP_KEY_TOOL,
-    DEP_KEY_VERSION,
-    DEP_XML_PACKAGE_REF,
-    DEPENDENCY_FILES,
-    DICT_KEY_RESULTS,
-    DIR_MTIMES_FILENAME,
-    DIST_DIR,
-    DOC_PROMPT_PREFIX,
-    DOCSTRING_STRIP_CHARS,
-    DUP_QN_MARKER,
-    EMBEDDING_CACHE_FILENAME,
-    EMBEDDING_DEFAULT_BATCH_SIZE,
-    EMPTY_PARENS,
-    ENCODING_UTF8,
-    ENTITY_CLASS,
-    ENTITY_CONSTRUCTOR,
-    ENTITY_FUNCTION,
-    ENTITY_METHOD,
-    ERR_SUBSTR_ALREADY_EXISTS,
-    ERR_SUBSTR_CONSTRAINT,
-    ESCAPE_CHAR,
-    EXCLUDED_DEPENDENCY_NAMES,
-    FIELD_MACRO,
-    FQN_C_FUNCTION_TYPES,
-    FQN_C_SCOPE_TYPES,
-    FQN_CPP_FUNCTION_TYPES,
-    FQN_CPP_SCOPE_TYPES,
-    FQN_GO_FUNCTION_TYPES,
-    FQN_GO_SCOPE_TYPES,
-    FQN_JAVA_FUNCTION_TYPES,
-    FQN_JAVA_SCOPE_TYPES,
-    FQN_JS_FUNCTION_TYPES,
-    FQN_JS_SCOPE_TYPES,
-    FQN_LUA_FUNCTION_TYPES,
-    FQN_LUA_SCOPE_TYPES,
-    FQN_PHP_FUNCTION_TYPES,
-    FQN_PHP_SCOPE_TYPES,
-    FQN_PY_FUNCTION_TYPES,
-    FQN_PY_SCOPE_TYPES,
-    FQN_RS_FUNCTION_TYPES,
-    FQN_RS_SCOPE_TYPES,
-    FQN_SCALA_FUNCTION_TYPES,
-    FQN_SCALA_SCOPE_TYPES,
-    FQN_TS_FUNCTION_TYPES,
-    FQN_TS_SCOPE_TYPES,
-    FUNCTION_NODES_BASIC,
-    FUNCTION_NODES_GENERATOR,
-    FUNCTION_NODES_LAMBDA,
-    FUNCTION_NODES_METHOD,
-    FUNCTION_NODES_TEMPLATE,
-    GEMFILE_GEM_PREFIX,
-    GIT_DIR_NAME,
-    GITWILDMATCH_STYLE,
-    GLOB_ALL,
-    GO_KEYWORD_MODULE,
-    GO_MOD_COMMENT_PREFIX,
-    GO_TEST_FILE_SUFFIX,
-    GOMOD_BLOCK_END,
-    GOMOD_COMMENT_PREFIX,
-    GOMOD_REQUIRE_BLOCK_START,
-    GOMOD_REQUIRE_LINE_PREFIX,
-    GOOGLE_CLOUD_SCOPE,
-    GUARD_INHERITED_METHOD,
-    HASH_CACHE_FILENAME,
-    HEADER_CONTENT_TYPE,
-    HEALTH_CHECK_API_KEY_CONFIGURED,
-    HEALTH_CHECK_API_KEY_MISSING_MSG,
-    HEALTH_CHECK_API_KEY_NOT_CONFIGURED,
-    HEALTH_CHECK_API_KEY_NOT_SET,
-    HEALTH_CHECK_API_KEY_SET,
-    HEALTH_CHECK_DOCKER_EXIT_CODE,
-    HEALTH_CHECK_DOCKER_FAILED_MSG,
-    HEALTH_CHECK_DOCKER_NOT_IN_PATH,
-    HEALTH_CHECK_DOCKER_NOT_INSTALLED_MSG,
-    HEALTH_CHECK_DOCKER_NOT_RESPONDING_MSG,
-    HEALTH_CHECK_DOCKER_NOT_RUNNING,
-    HEALTH_CHECK_DOCKER_RUNNING,
-    HEALTH_CHECK_DOCKER_RUNNING_MSG,
-    HEALTH_CHECK_DOCKER_TIMEOUT_ERROR,
-    HEALTH_CHECK_DOCKER_TIMEOUT_MSG,
-    HEALTH_CHECK_EXTERNAL_TOOLS,
-    HEALTH_CHECK_GRAPH_INTEGRITY_ERROR_MSG,
-    HEALTH_CHECK_GRAPH_INTEGRITY_FAILED,
-    HEALTH_CHECK_GRAPH_INTEGRITY_OK,
-    HEALTH_CHECK_GRAPH_INTEGRITY_OK_MSG,
-    HEALTH_CHECK_GRAPH_INTEGRITY_SEPARATOR,
-    HEALTH_CHECK_GRAPH_INTEGRITY_VIOLATIONS_MSG,
-    HEALTH_CHECK_MEMGRAPH_CONNECTED_MSG,
-    HEALTH_CHECK_MEMGRAPH_CONNECTION_FAILED_MSG,
-    HEALTH_CHECK_MEMGRAPH_ERROR,
-    HEALTH_CHECK_MEMGRAPH_FAILED,
-    HEALTH_CHECK_MEMGRAPH_QUERY,
-    HEALTH_CHECK_MEMGRAPH_SUCCESSFUL,
-    HEALTH_CHECK_MEMGRAPH_UNEXPECTED_FAILURE_MSG,
-    HEALTH_CHECK_TOOL_FAILED_MSG,
-    HEALTH_CHECK_TOOL_INSTALLED,
-    HEALTH_CHECK_TOOL_INSTALLED_MSG,
-    HEALTH_CHECK_TOOL_NOT_IN_PATH_MSG,
-    HEALTH_CHECK_TOOL_NOT_INSTALLED,
-    HEALTH_CHECK_TOOL_TIMEOUT_ERROR,
-    HEALTH_CHECK_TOOL_TIMEOUT_MSG,
-    HEALTH_CHECK_TOOLS,
-    HIGHER_ORDER_BUILTINS,
-    HTTP_OK,
-    IGNORE_PATTERNS,
-    IGNORE_SUFFIXES,
-    IIFE_ARROW_PREFIX,
-    IIFE_FUNC_PREFIX,
-    IMPORT_CACHE_DIR,
-    IMPORT_CACHE_FILE,
-    IMPORT_CACHE_KEY,
-    IMPORT_CACHE_TTL,
-    IMPORT_DEFAULT_SUFFIX,
-    IMPORT_IMPORT,
-    IMPORT_MODULE_LABEL,
-    IMPORT_NODES_FROM,
-    IMPORT_NODES_INCLUDE,
-    IMPORT_NODES_MODULE,
-    IMPORT_NODES_STANDARD,
-    IMPORT_PCALL,
-    IMPORT_QUALIFIED_NAME,
-    IMPORT_RELATIONSHIP,
-    IMPORT_REQUIRE,
-    IMPORT_STD_PREFIX,
-    IMPORT_TIMESTAMPS_KEY,
-    INDEX_LUA_INIT,
-    INIT_PY,
-    INLINE_MODULE_PATH_PREFIX,
-    JAVA_ARRAY_SUFFIX,
-    JAVA_CALLABLE_ENTITY_TYPES,
-    JAVA_CLASS_MODIFIERS,
-    JAVA_CLASS_NODE_TYPES,
-    JAVA_CREATE_PATTERN,
-    JAVA_DECLARATION_SUFFIX,
-    JAVA_FIELD_MODIFIERS,
-    JAVA_GETTER_PATTERN,
-    JAVA_HAS_PATTERN,
-    JAVA_HEURISTIC_ORDER,
-    JAVA_HEURISTIC_USER,
-    JAVA_ID_PATTERN,
-    JAVA_IS_PATTERN,
-    JAVA_JVM_LANGUAGES,
-    JAVA_KEYWORD_SUPER,
-    JAVA_KEYWORD_THIS,
-    JAVA_LANG_CLASS_NAMES,
-    JAVA_LANG_PREFIX,
-    JAVA_LENGTH_PATTERN,
-    JAVA_MAIN_METHOD_NAME,
-    JAVA_MAIN_PARAM_ARRAY,
-    JAVA_MAIN_PARAM_TYPE,
-    JAVA_MAIN_PARAM_VARARGS,
-    JAVA_METHOD_MODIFIERS,
-    JAVA_METHOD_NODE_TYPES,
-    JAVA_MODIFIER_ABSTRACT,
-    JAVA_MODIFIER_FINAL,
-    JAVA_MODIFIER_PRIVATE,
-    JAVA_MODIFIER_PROTECTED,
-    JAVA_MODIFIER_PUBLIC,
-    JAVA_MODIFIER_STATIC,
-    JAVA_MODIFIER_SYNCHRONIZED,
-    JAVA_MODIFIER_TRANSIENT,
-    JAVA_MODIFIER_VOLATILE,
-    JAVA_NAME_PATTERN,
-    JAVA_NEW_PATTERN,
-    JAVA_ORDER_PATTERN,
-    JAVA_PATH_JAVA,
-    JAVA_PATH_KOTLIN,
-    JAVA_PATH_MAIN,
-    JAVA_PATH_SCALA,
-    JAVA_PATH_SRC,
-    JAVA_PATH_TEST,
-    JAVA_PRIMITIVE_TYPES,
-    JAVA_SERIALIZATION_METHOD_NAMES,
-    JAVA_SIZE_PATTERN,
-    JAVA_SRC_FOLDERS,
-    JAVA_STDLIB_CLASSES,
-    JAVA_STDLIB_PREFIXES,
-    JAVA_SUFFIX_BUILDER,
-    JAVA_SUFFIX_ERROR,
-    JAVA_SUFFIX_EXCEPTION,
-    JAVA_SUFFIX_INTERFACE,
-    JAVA_TYPE_BOOLEAN,
-    JAVA_TYPE_CONSTRUCTOR,
-    JAVA_TYPE_DOUBLE,
-    JAVA_TYPE_INT,
-    JAVA_TYPE_LONG,
-    JAVA_TYPE_METHOD,
-    JAVA_TYPE_OBJECT,
-    JAVA_TYPE_STRING,
-    JAVA_TYPE_STRING_FQN,
-    JAVA_USER_PATTERN,
-    JAVA_VISIBILITY_PACKAGE,
-    JAVA_VISIBILITY_PRIVATE,
-    JAVA_VISIBILITY_PROTECTED,
-    JAVA_VISIBILITY_PUBLIC,
-    JAVA_WRAPPER_TYPES,
-    JS_ARROW_NAME_CLIMB_BOUNDARY,
-    JS_ASSIGNMENT_ARROW_QUERY,
-    JS_ASSIGNMENT_FUNCTION_QUERY,
-    JS_BUILTIN_PATTERNS,
-    JS_BUILTIN_TYPES,
-    JS_CALL_RESULT_VALUE_TYPES,
-    JS_COMMONJS_DESTRUCTURE_QUERY,
-    JS_COMMONJS_EXPORTS_FUNCTION_QUERY,
-    JS_COMMONJS_MODULE_EXPORTS_QUERY,
-    JS_CREATE_METHOD,
-    JS_ES6_EXPORT_CONST_QUERY,
-    JS_ES6_EXPORT_FUNCTION_QUERY,
-    JS_EXPORT_TYPE_COMMONJS,
-    JS_EXPORT_TYPE_COMMONJS_MODULE,
-    JS_EXPORT_TYPE_ES6_FUNCTION,
-    JS_EXPORT_TYPE_ES6_FUNCTION_DECL,
-    JS_EXPORTS_KEYWORD,
-    JS_FUNCTION_PROTOTYPE_METHODS,
-    JS_FUNCTION_PROTOTYPE_SUFFIXES,
-    JS_GLOBAL_CLASS_NAMES,
-    JS_METHOD_APPLY,
-    JS_METHOD_BIND,
-    JS_METHOD_CALL,
-    JS_METHOD_DEF_QUERY,
-    JS_MODULE_KEYWORD,
-    JS_MODULE_RECEIVERS,
-    JS_NAME_NODE_TYPES,
-    JS_OBJECT_ARROW_QUERY,
-    JS_OBJECT_METHOD_QUERY,
-    JS_OBJECT_NAME,
-    JS_PROTOTYPE_INHERITANCE_QUERY,
-    JS_PROTOTYPE_KEYWORD,
-    JS_PROTOTYPE_METHOD_QUERY,
-    JS_REQUIRE_KEYWORD,
-    JS_SUFFIX_APPLY,
-    JS_SUFFIX_BIND,
-    JS_SUFFIX_CALL,
-    JS_TS_ALL_EXTENSIONS,
-    JS_TS_LANGUAGES,
-    JS_TS_PARENT_REF_TYPES,
-    JSON_EMPTY_OBJECT,
-    JSON_INDENT,
-    JSON_KEY_ENTITY_TYPE,
-    JSON_KEY_HAS_ENTITY,
-    KEY_IS_MACRO,
-    KEY_OVERRIDES_EXTERNAL,
-    KEYWORD_CONSTRUCTOR,
-    KEYWORD_SELF,
-    KEYWORD_SUPER,
-    LANG_CALL_KEYWORD_EXCLUDE,
-    LANG_CALL_KEYWORDS,
-    LANG_CLASS_KEYWORDS,
-    LANG_CONFIG_FILE,
-    LANG_DEFAULT_GRAMMAR_URL,
-    LANG_ELLIPSIS,
-    LANG_ERR_CONFIG_NOT_FOUND,
-    LANG_ERR_CUSTOM_URL_HINT,
-    LANG_ERR_GIT,
-    LANG_ERR_MANUAL_REMOVE_HINT,
-    LANG_ERR_MISSING_ARGS,
-    LANG_ERR_NO_GRAMMARS_WARNING,
-    LANG_ERR_NODE_TYPES_WARNING,
-    LANG_ERR_PARSE_NODE_TYPES,
-    LANG_ERR_REINSTALL_FAILED,
-    LANG_ERR_REMOVE_CONFIG,
-    LANG_ERR_REMOVE_SUBMODULE,
-    LANG_ERR_REPO_NOT_FOUND,
-    LANG_ERR_TREE_SITTER_JSON_WARNING,
-    LANG_ERR_UPDATE_CONFIG,
-    LANG_EXCLUSION_KEYWORDS,
-    LANG_FALLBACK_MANUAL_ADD,
-    LANG_FALLBACK_METHOD_NODE,
-    LANG_FUNCTION_KEYWORDS,
-    LANG_GIT_MODULES_PATH,
-    LANG_GIT_SUFFIX,
-    LANG_GITMODULES_FILE,
-    LANG_GITMODULES_REGEX,
-    LANG_GRAMMARS_DIR,
-    LANG_MODULE_KEYWORDS,
-    LANG_MSG_ADDING_SUBMODULE,
-    LANG_MSG_AUTO_DETECTED_EXT,
-    LANG_MSG_AUTO_DETECTED_LANG,
-    LANG_MSG_AVAILABLE_LANGS,
-    LANG_MSG_CALLS,
-    LANG_MSG_CATEGORY_FORMAT,
-    LANG_MSG_CLASSES,
-    LANG_MSG_CLEANED_MODULES,
-    LANG_MSG_CLEANUP_CANCELLED,
-    LANG_MSG_CLEANUP_COMPLETE,
-    LANG_MSG_COMMON_ISSUES,
-    LANG_MSG_CUSTOM_URL_WARNING,
-    LANG_MSG_EDIT_HINT,
-    LANG_MSG_FOUND_NODE_TYPES,
-    LANG_MSG_FOUND_ORPHANS,
-    LANG_MSG_FUNCTIONS,
-    LANG_MSG_ISSUE_CALL_TYPES,
-    LANG_MSG_ISSUE_CLASS_TYPES,
-    LANG_MSG_ISSUE_MISCLASSIFIED,
-    LANG_MSG_ISSUE_MISSING,
-    LANG_MSG_KEEPING_SUBMODULE,
-    LANG_MSG_LANG_ADDED,
-    LANG_MSG_LANG_NOT_FOUND,
-    LANG_MSG_LANG_REMOVED,
-    LANG_MSG_LIST_HINT,
-    LANG_MSG_MAPPED_CATEGORIES,
-    LANG_MSG_MODULES,
-    LANG_MSG_NO_GITMODULES,
-    LANG_MSG_NO_MODULES_DIR,
-    LANG_MSG_NO_ORPHANS,
-    LANG_MSG_NO_SUBMODULE,
-    LANG_MSG_READDING_SUBMODULE,
-    LANG_MSG_REINSTALL_SUCCESS,
-    LANG_MSG_REMOVED_FROM_CONFIG,
-    LANG_MSG_REMOVED_ORPHAN,
-    LANG_MSG_REMOVING_ENTRY,
-    LANG_MSG_REMOVING_SUBMODULE,
-    LANG_MSG_REVIEW_HINT,
-    LANG_MSG_REVIEW_PROMPT,
-    LANG_MSG_SEMANTIC_CATEGORIES,
-    LANG_MSG_SUBMODULE_EXISTS,
-    LANG_MSG_SUBMODULE_REMOVED,
-    LANG_MSG_SUBMODULE_SUCCESS,
-    LANG_MSG_UPDATED_CONFIG,
-    LANG_MSG_USING_DEFAULT_URL,
-    LANG_MSG_USING_LANG_NAME,
-    LANG_NODE_TYPES_JSON,
-    LANG_PROMPT_CALLS,
-    LANG_PROMPT_CLASSES,
-    LANG_PROMPT_COMMON_NAME,
-    LANG_PROMPT_CONTINUE,
-    LANG_PROMPT_FUNCTIONS,
-    LANG_PROMPT_LANGUAGE_NAME,
-    LANG_PROMPT_MODULES,
-    LANG_PROMPT_REMOVE_ORPHANS,
-    LANG_SRC_DIR,
-    LANG_TABLE_COL_CALL_TYPES,
-    LANG_TABLE_COL_CLASS_TYPES,
-    LANG_TABLE_COL_FUNCTION_TYPES,
-    LANG_TABLE_COL_LANGUAGE,
-    LANG_TABLE_PLACEHOLDER,
-    LANG_TABLE_TITLE,
-    LANG_TREE_SITTER_JSON,
-    LANG_TREE_SITTER_URL_MARKER,
-    LUA_DEFAULT_VAR_TYPES,
-    LUA_METHOD_SEPARATOR,
-    LUA_STATEMENT_SUFFIX,
-    LUA_STDLIB_MODULES,
-    LUA_STRING_TYPES,
-    MARKDOWN_FENCE,
-    MARKDOWN_FENCE_DIFF,
-    MCP_ASK_AGENT_ERROR,
-    MCP_CONTENT_TYPE_TEXT,
-    MCP_DEFAULT_DIRECTORY,
-    MCP_INDEX_ERROR,
-    MCP_INDEX_SUCCESS,
-    MCP_INDEX_SUCCESS_PROJECT,
-    MCP_JSON_INDENT,
-    MCP_KEY_ERROR,
-    MCP_KEY_ERROR_MESSAGE,
-    MCP_KEY_FOUND,
-    MCP_KEY_QUERY_USED,
-    MCP_KEY_RESULTS,
-    MCP_KEY_SUMMARY,
-    MCP_LOG_FORMAT,
-    MCP_LOG_LEVEL_INFO,
-    MCP_NOT_AVAILABLE,
-    MCP_PAGINATION_HEADER,
-    MCP_PROJECT_DELETED,
-    MCP_SEMANTIC_NOT_AVAILABLE_RESPONSE,
-    MCP_SERVER_NAME,
-    MCP_TOOL_EXEC_ERROR,
-    MCP_TOOL_NAME_QUERY,
-    MCP_UNKNOWN_TOOL_ERROR,
-    MCP_UPDATE_ERROR,
-    MCP_UPDATE_SUCCESS,
-    MCP_WIPE_CANCELLED,
-    MCP_WIPE_ERROR,
-    MCP_WIPE_SUCCESS,
-    MCP_WRITE_SUCCESS,
-    MIME_TYPE_DEFAULT,
-    ML_DEPENDENCIES,
-    MOD_RS,
-    MODEL_CONTEXT_WINDOWS,
-    MODULE_INDEX_FILE_STEMS,
-    MODULE_QDRANT_CLIENT,
-    MODULE_TORCH,
-    MODULE_TRANSFORMERS,
-    OLLAMA_HEALTH_PATH,
-    OPENAI_DEFAULT_ENDPOINT,
-    OPERATOR_PREFIX,
-    PACKAGE_NAME,
-    PATH_CURRENT_DIR,
-    PATH_PARENT_DIR,
-    PKG_CARGO_TOML,
-    PKG_CMAKE_LISTS,
-    PKG_CONANFILE,
-    PKG_INIT_PY,
-    PKG_MAKEFILE,
-    PKG_VCXPROJ_GLOB,
-    PREFIX_ANONYMOUS,
-    PREFIX_ARROW,
-    PREFIX_FUNC,
-    PREFIX_IIFE,
-    PREFIX_IIFE_DIRECT,
-    PREFIX_LAMBDA,
-    PROPERTY_DECORATORS,
-    PROTOBUF_INDEX_FILE,
-    PROTOBUF_NODES_FILE,
-    PROTOBUF_RELS_FILE,
-    PUNCTUATION_TYPES,
-    PY_ASSIGNMENT_QUERY,
-    PY_BUILTIN_GETATTR,
-    PY_BUILTIN_LEN,
-    PY_CLASS_REPOSITORY,
-    PY_CLS_PREFIX,
-    PY_DUNDER_BOOL,
-    PY_DUNDER_CONTAINS,
-    PY_DUNDER_GETITEM,
-    PY_DUNDER_LEN,
-    PY_DUNDER_SETITEM,
-    PY_KEYWORD_CLS,
-    PY_KEYWORD_SELF,
-    PY_METHOD_CREATE,
-    PY_METHOD_INIT,
-    PY_MODELS_BASE_PATH,
-    PY_NAME_DUNDER,
-    PY_NAME_UNDERSCORE,
-    PY_NONE,
-    PY_OP_IN,
-    PY_OPERAND_REJECT_CHARS,
-    PY_PROTOCOL,
-    PY_RETURN_QUERY,
-    PY_SCORE_CONTAINS_BASE,
-    PY_SCORE_EXACT_MATCH,
-    PY_SCORE_SUFFIX_MATCH,
-    PY_SELF_PREFIX,
-    PY_UNION_SEPARATOR,
-    PY_VAR_PATTERN_ALL,
-    PY_VAR_SUFFIX_PLURAL,
-    PYINSTALLER_ARG_CLEAN,
-    PYINSTALLER_ARG_COLLECT_ALL,
-    PYINSTALLER_ARG_COLLECT_DATA,
-    PYINSTALLER_ARG_COPY_METADATA,
-    PYINSTALLER_ARG_EXCLUDE_MODULE,
-    PYINSTALLER_ARG_HIDDEN_IMPORT,
-    PYINSTALLER_ARG_NAME,
-    PYINSTALLER_ARG_NOCONFIRM,
-    PYINSTALLER_ARG_ONEFILE,
-    PYINSTALLER_CMD,
-    PYINSTALLER_ENTRY_POINT,
-    PYINSTALLER_EXCLUDED_MODULES,
-    PYINSTALLER_PACKAGES,
-    PYPROJECT_KEY_PACKAGE_DIR,
-    PYPROJECT_KEY_SETUPTOOLS,
-    PYPROJECT_KEY_TOOL,
-    PYPROJECT_PATH,
-    QUOTE_CHARS,
-    REGEX_FINAL_METHOD_CAPTURE,
-    REGEX_METHOD_CHAIN_SUFFIX,
-    ROOT_DIR_KEY,
-    RS_CALL_OR_GENERIC_FN,
-    RS_COMMENT_TYPES,
-    RS_DEREF_WRAPPERS,
-    RS_ENCODING_UTF8,
-    RS_FIELD_ARGUMENT,
-    RS_GUARD_ACCESSORS,
-    RS_GUARD_WRAPPERS,
-    RS_IDENT_NODE_TYPES,
-    RS_IDENT_OR_SELF,
-    RS_IDENTIFIER_TYPES,
-    RS_IDENTITY_METHODS,
-    RS_MACRO_EXPORT_ATTR,
-    RS_MACRO_RECEIVER_TYPES,
-    RS_PATH_KEYWORDS,
-    RS_RESULT_WRAPPERS,
-    RS_RETURN_STRIP_WRAPPERS,
-    RS_RETURN_TYPE_NODE_TYPES,
-    RS_SCOPED_TYPES,
-    RS_SELF_TYPE,
-    RS_TYPE_NODE_TYPES,
-    RS_USE_LIST_DELIMITERS,
-    RS_WILDCARD_PREFIX,
-    RUST_CRATE_PREFIX,
-    SCHEMA_OPTIONAL_SUFFIX,
-    SCHEMA_PROPS_BRACES,
-    SEMANTIC_BATCH_SIZE,
-    SEMANTIC_DEPENDENCIES,
-    SEMANTIC_TYPE_UNKNOWN,
-    SEPARATOR_COLON,
-    SEPARATOR_COMMA,
-    SEPARATOR_COMMA_SPACE,
-    SEPARATOR_DOT,
-    SEPARATOR_DOUBLE_COLON,
-    SEPARATOR_PROTOTYPE,
-    SEPARATOR_SLASH,
-    SHIFT_TAB_ESCAPE,
-    SPEC_C_CALL_TYPES,
-    SPEC_C_CLASS_TYPES,
-    SPEC_C_FUNCTION_TYPES,
-    SPEC_C_MODULE_TYPES,
-    SPEC_C_PACKAGE_INDICATORS,
-    SPEC_CPP_CALL_TYPES,
-    SPEC_CPP_CLASS_TYPES,
-    SPEC_CPP_FUNCTION_TYPES,
-    SPEC_CPP_MODULE_TYPES,
-    SPEC_CPP_PACKAGE_INDICATORS,
-    SPEC_GO_CALL_TYPES,
-    SPEC_GO_CLASS_TYPES,
-    SPEC_GO_FUNCTION_TYPES,
-    SPEC_GO_IMPORT_TYPES,
-    SPEC_GO_MODULE_TYPES,
-    SPEC_JAVA_CALL_TYPES,
-    SPEC_JAVA_CLASS_TYPES,
-    SPEC_JAVA_FUNCTION_TYPES,
-    SPEC_JAVA_IMPORT_TYPES,
-    SPEC_JAVA_MODULE_TYPES,
-    SPEC_JS_CALL_TYPES,
-    SPEC_JS_MODULE_TYPES,
-    SPEC_LUA_CALL_TYPES,
-    SPEC_LUA_CLASS_TYPES,
-    SPEC_LUA_FUNCTION_TYPES,
-    SPEC_LUA_IMPORT_TYPES,
-    SPEC_LUA_MODULE_TYPES,
-    SPEC_PHP_CALL_TYPES,
-    SPEC_PHP_CLASS_TYPES,
-    SPEC_PHP_FUNCTION_TYPES,
-    SPEC_PHP_IMPORT_FROM_TYPES,
-    SPEC_PHP_IMPORT_TYPES,
-    SPEC_PHP_MODULE_TYPES,
-    SPEC_PY_CALL_TYPES,
-    SPEC_PY_CLASS_TYPES,
-    SPEC_PY_FUNCTION_TYPES,
-    SPEC_PY_IMPORT_FROM_TYPES,
-    SPEC_PY_IMPORT_TYPES,
-    SPEC_PY_MODULE_TYPES,
-    SPEC_PY_PACKAGE_INDICATORS,
-    SPEC_RS_CALL_TYPES,
-    SPEC_RS_CLASS_TYPES,
-    SPEC_RS_FUNCTION_TYPES,
-    SPEC_RS_IMPORT_FROM_TYPES,
-    SPEC_RS_IMPORT_TYPES,
-    SPEC_RS_MODULE_TYPES,
-    SPEC_RS_PACKAGE_INDICATORS,
-    SPEC_SCALA_CALL_TYPES,
-    SPEC_SCALA_CLASS_TYPES,
-    SPEC_SCALA_FUNCTION_TYPES,
-    SPEC_SCALA_IMPORT_TYPES,
-    SPEC_SCALA_MODULE_TYPES,
-    STR_NONE,
-    TEXT_UNKNOWN,
-    TIKTOKEN_ENCODING,
-    TMP_EXTENSION,
-    TOML_KEY_OPTIONAL_DEPS,
-    TOML_KEY_PROJECT,
-    TREESITTER_EXTRA_KEY,
-    TREESITTER_PKG_PREFIX,
-    TRIPLE_QUOTES,
-    TS_ABSTRACT_CLASS_DECLARATION,
-    TS_ACCESS_SPECIFIER,
-    TS_ACCESSIBILITY_MODIFIER,
-    TS_ALIASED_IMPORT,
-    TS_ANNOTATION,
-    TS_ANNOTATION_TYPE_DECLARATION,
-    TS_ARGUMENTS,
-    TS_ARRAY,
-    TS_ARRAY_CREATION_EXPRESSION,
-    TS_ARROW_FUNCTION,
-    TS_AS_EXPRESSION,
-    TS_ASSIGNMENT_EXPRESSION,
-    TS_ASTERISK,
-    TS_ATTRIBUTE,
-    TS_BASE_CLASS_CLAUSE,
-    TS_BINARY_EXPRESSION,
-    TS_CALL_EXPRESSION,
-    TS_CAST_WRAPPER_TYPES,
-    TS_CLASS_BODY,
-    TS_CLASS_DECLARATION,
-    TS_CLASS_HERITAGE,
-    TS_CONSTRUCTOR_DECLARATION,
-    TS_CPP_BINARY_EXPRESSION,
-    TS_CPP_CALL_EXPRESSION,
-    TS_CPP_DECLARATION,
-    TS_CPP_DELETE_EXPRESSION,
-    TS_CPP_FIELD_DECLARATION,
-    TS_CPP_FIELD_EXPRESSION,
-    TS_CPP_FUNCTION_DECLARATOR,
-    TS_CPP_FUNCTION_DEFINITION,
-    TS_CPP_LAMBDA_EXPRESSION,
-    TS_CPP_LINKAGE_SPECIFICATION,
-    TS_CPP_NEW_EXPRESSION,
-    TS_CPP_SUBSCRIPT_EXPRESSION,
-    TS_CPP_TEMPLATE_DECLARATION,
-    TS_CPP_TEMPLATE_PARAMETER_LIST,
-    TS_CPP_TRANSLATION_UNIT,
-    TS_CPP_UNARY_EXPRESSION,
-    TS_CPP_UPDATE_EXPRESSION,
-    TS_DECIMAL_FLOATING_POINT_LITERAL,
-    TS_DECLARATION,
-    TS_DECORATOR,
-    TS_DOT_INDEX_EXPRESSION,
-    TS_DOTTED_NAME,
-    TS_ENHANCED_FOR_STATEMENT,
-    TS_ENUM_CLASS_SPECIFIER,
-    TS_ENUM_DECLARATION,
-    TS_ENUM_SPECIFIER,
-    TS_ERROR,
-    TS_EXPORT_CLAUSE,
-    TS_EXPORT_DEFAULT,
-    TS_EXPORT_SPECIFIER,
-    TS_EXPORT_STATEMENT,
-    TS_EXPRESSION_STATEMENT,
-    TS_EXTENDS,
-    TS_EXTENDS_CLAUSE,
-    TS_EXTENDS_TYPE_CLAUSE,
-    TS_FALSE,
-    TS_FIELD_ACCESS,
-    TS_FIELD_ARGUMENT,
-    TS_FIELD_ARGUMENTS,
-    TS_FIELD_BODY,
-    TS_FIELD_CONDITION,
-    TS_FIELD_DECLARATION,
-    TS_FIELD_DECLARATOR,
-    TS_FIELD_FUNCTION,
-    TS_FIELD_INTERFACES,
-    TS_FIELD_LEFT,
-    TS_FIELD_NAME,
-    TS_FIELD_OBJECT,
-    TS_FIELD_OPERATORS,
-    TS_FIELD_PARAMETER,
-    TS_FIELD_PARAMETERS,
-    TS_FIELD_PATTERN,
-    TS_FIELD_RIGHT,
-    TS_FIELD_SUPERCLASS,
-    TS_FIELD_TYPE,
-    TS_FIELD_TYPE_PARAMETERS,
-    TS_FORMAL_PARAMETER,
-    TS_FUNCTION_DECLARATION,
-    TS_FUNCTION_EXPRESSION,
-    TS_FUNCTION_SIGNATURE,
-    TS_GENERATOR_FUNCTION_DECLARATION,
-    TS_GENERIC_FUNCTION,
-    TS_GENERIC_TYPE,
-    TS_GO_ASSIGNMENT_STATEMENT,
-    TS_GO_CALL_EXPRESSION,
-    TS_GO_COMPOSITE_LITERAL,
-    TS_GO_CONTAINER_TYPES,
-    TS_GO_EXPRESSION_LIST,
-    TS_GO_FIELD_DECLARATION,
-    TS_GO_FIELD_DECLARATION_LIST,
-    TS_GO_FIELD_IDENTIFIER,
-    TS_GO_FUNC_LITERAL,
-    TS_GO_FUNCTION_DECLARATION,
-    TS_GO_IMPORT_DECLARATION,
-    TS_GO_INTERFACE_TYPE,
-    TS_GO_KEYED_ELEMENT,
-    TS_GO_LITERAL_ELEMENT,
-    TS_GO_LITERAL_VALUE,
-    TS_GO_METHOD_DECLARATION,
-    TS_GO_PARAMETER_DECLARATION,
-    TS_GO_PARAMETER_LIST,
-    TS_GO_POINTER_TYPE,
-    TS_GO_SELECTOR_EXPRESSION,
-    TS_GO_SHORT_VAR_DECLARATION,
-    TS_GO_SOURCE_FILE,
-    TS_GO_STRUCT_TYPE,
-    TS_GO_TYPE_ALIAS,
-    TS_GO_TYPE_ASSERTION_EXPRESSION,
-    TS_GO_TYPE_DECLARATION,
-    TS_GO_TYPE_SPEC,
-    TS_GO_UNARY_EXPRESSION,
-    TS_GO_VAR_DECLARATION,
-    TS_GO_VAR_SPEC,
-    TS_IDENTIFIER,
-    TS_IMPL_ITEM,
-    TS_IMPLEMENTS_CLAUSE,
-    TS_IMPORT_CLAUSE,
-    TS_IMPORT_DECLARATION,
-    TS_IMPORT_FROM_STATEMENT,
-    TS_IMPORT_PREFIX,
-    TS_IMPORT_SPEC,
-    TS_IMPORT_SPEC_LIST,
-    TS_IMPORT_SPECIFIER,
-    TS_IMPORT_STATEMENT,
-    TS_INTEGER_LITERAL,
-    TS_INTERFACE_DECLARATION,
-    TS_INTERNAL_MODULE,
-    TS_INTERPRETED_STRING_LITERAL,
-    TS_JAVA_ANNOTATION_TYPE_DECLARATION,
-    TS_JAVA_CAST_EXPRESSION,
-    TS_JAVA_EXTENDS_INTERFACES,
-    TS_JAVA_METHOD_INVOCATION,
-    TS_JSX_EXPRESSION,
-    TS_JSX_OPENING_ELEMENT,
-    TS_JSX_SELF_CLOSING_ELEMENT,
-    TS_LEXICAL_DECLARATION,
-    TS_LOCAL_VARIABLE_DECLARATION,
-    TS_LUA_ASSIGNMENT_STATEMENT,
-    TS_LUA_CHUNK,
-    TS_LUA_EXPRESSION_LIST,
-    TS_LUA_FUNCTION_CALL,
-    TS_LUA_FUNCTION_DECLARATION,
-    TS_LUA_FUNCTION_DEFINITION,
-    TS_LUA_IDENTIFIER,
-    TS_LUA_LOCAL_STATEMENT,
-    TS_LUA_METHOD_INDEX_EXPRESSION,
-    TS_LUA_VARIABLE_DECLARATION,
-    TS_LUA_VARIABLE_LIST,
-    TS_MARKER_ANNOTATION,
-    TS_MEMBER_EXPRESSION,
-    TS_METHOD_DECLARATION,
-    TS_METHOD_DEFINITION,
-    TS_METHOD_INVOCATION,
-    TS_MODIFIERS,
-    TS_MODULE,
-    TS_MODULE_DECLARATION,
-    TS_NAMED_IMPORTS,
-    TS_NAMESPACE_DEFINITION,
-    TS_NAMESPACE_IMPORT,
-    TS_NEW_EXPRESSION,
-    TS_NON_NULL_EXPRESSION,
-    TS_OBJECT,
-    TS_OBJECT_CREATION_EXPRESSION,
-    TS_OBJECT_PATTERN,
-    TS_OPTIONAL_PARAMETER,
-    TS_PACKAGE_DECLARATION,
-    TS_PACKAGE_IDENTIFIER,
-    TS_PAIR,
-    TS_PAIR_PATTERN,
-    TS_PARENTHESIZED_EXPRESSION,
-    TS_PHP_ANONYMOUS_FUNCTION,
-    TS_PHP_ARROW_FUNCTION,
-    TS_PHP_ATTRIBUTE,
-    TS_PHP_ATTRIBUTE_GROUP,
-    TS_PHP_ATTRIBUTE_LIST,
-    TS_PHP_BASE_CLAUSE,
-    TS_PHP_CLASS_INTERFACE_CLAUSE,
-    TS_PHP_FUNCTION,
-    TS_PHP_FUNCTION_CALL_EXPRESSION,
-    TS_PHP_FUNCTION_DEFINITION,
-    TS_PHP_FUNCTION_STATIC_DECLARATION,
-    TS_PHP_INCLUDE_EXPRESSION,
-    TS_PHP_INCLUDE_ONCE_EXPRESSION,
-    TS_PHP_MEMBER_CALL_EXPRESSION,
-    TS_PHP_METHOD_DECLARATION,
-    TS_PHP_NAME,
-    TS_PHP_NAMESPACE_DEFINITION,
-    TS_PHP_NAMESPACE_USE_CLAUSE,
-    TS_PHP_NAMESPACE_USE_DECLARATION,
-    TS_PHP_NULLSAFE_MEMBER_CALL_EXPRESSION,
-    TS_PHP_OBJECT_CREATION_EXPRESSION,
-    TS_PHP_QUALIFIED_NAME,
-    TS_PHP_REQUIRE_EXPRESSION,
-    TS_PHP_REQUIRE_ONCE_EXPRESSION,
-    TS_PHP_SCOPED_CALL_EXPRESSION,
-    TS_PHP_TRAIT_DECLARATION,
-    TS_PHP_USE_DECLARATION,
-    TS_PHP_VISIBILITY_MODIFIER,
-    TS_PREPROC_INCLUDE,
-    TS_PRIVATE,
-    TS_PRIVATE_PROPERTY_IDENTIFIER,
-    TS_PROGRAM,
-    TS_PROPERTY_IDENTIFIER,
-    TS_PY_ASSIGNMENT,
-    TS_PY_ATTRIBUTE,
-    TS_PY_BLOCK,
-    TS_PY_BOOLEAN_OPERATOR,
-    TS_PY_CALL,
-    TS_PY_CLASS_DEFINITION,
-    TS_PY_COMPARISON_OPERATOR,
-    TS_PY_CONDITIONAL_EXPRESSION,
-    TS_PY_DECORATED_DEFINITION,
-    TS_PY_DECORATOR,
-    TS_PY_DEFAULT_PARAMETER,
-    TS_PY_DICTIONARY,
-    TS_PY_DICTIONARY_SPLAT_PATTERN,
-    TS_PY_ELIF_CLAUSE,
-    TS_PY_EXPRESSION_STATEMENT,
-    TS_PY_FOR_IN_CLAUSE,
-    TS_PY_FOR_STATEMENT,
-    TS_PY_FUNCTION_DEFINITION,
-    TS_PY_IDENTIFIER,
-    TS_PY_IF_STATEMENT,
-    TS_PY_IMPORT_FROM_STATEMENT,
-    TS_PY_IMPORT_STATEMENT,
-    TS_PY_KEYWORD,
-    TS_PY_KEYWORD_ARGUMENT,
-    TS_PY_LAMBDA,
-    TS_PY_LIST,
-    TS_PY_LIST_COMPREHENSION,
-    TS_PY_LIST_SPLAT_PATTERN,
-    TS_PY_MODULE,
-    TS_PY_NOT_OPERATOR,
-    TS_PY_PAIR,
-    TS_PY_RETURN,
-    TS_PY_RETURN_STATEMENT,
-    TS_PY_SET,
-    TS_PY_STRING,
-    TS_PY_STRING_CONTENT,
-    TS_PY_SUBSCRIPT,
-    TS_PY_TRY_STATEMENT,
-    TS_PY_TUPLE,
-    TS_PY_TYPED_DEFAULT_PARAMETER,
-    TS_PY_TYPED_PARAMETER,
-    TS_PY_WHILE_STATEMENT,
-    TS_PY_WITH_STATEMENT,
-    TS_RECORD_DECLARATION,
-    TS_RELATIVE_IMPORT,
-    TS_REQUIRED_PARAMETER,
-    TS_RETURN,
-    TS_RETURN_STATEMENT,
-    TS_RS_ATTRIBUTE_ITEM,
-    TS_RS_BLOCK_COMMENT,
-    TS_RS_CALL_EXPRESSION,
-    TS_RS_CLOSURE_EXPRESSION,
-    TS_RS_CRATE,
-    TS_RS_ENUM_ITEM,
-    TS_RS_EXTERN_CRATE_DECLARATION,
-    TS_RS_FIELD_DECLARATION,
-    TS_RS_FIELD_DECLARATION_LIST,
-    TS_RS_FIELD_EXPRESSION,
-    TS_RS_FIELD_IDENTIFIER,
-    TS_RS_FIELD_PATH,
-    TS_RS_FUNCTION_ITEM,
-    TS_RS_FUNCTION_SIGNATURE_ITEM,
-    TS_RS_IMPL_ITEM,
-    TS_RS_INNER_ATTRIBUTE_ITEM,
-    TS_RS_KEYWORD_AS,
-    TS_RS_LET_DECLARATION,
-    TS_RS_LINE_COMMENT,
-    TS_RS_MACRO_DEFINITION,
-    TS_RS_MACRO_INVOCATION,
-    TS_RS_MATCH_ARM,
-    TS_RS_MATCH_EXPRESSION,
-    TS_RS_MOD_ITEM,
-    TS_RS_PARAMETER,
-    TS_RS_POINTER_TYPE,
-    TS_RS_PRIMITIVE_TYPE,
-    TS_RS_REFERENCE_TYPE,
-    TS_RS_SCOPED_TYPE_IDENTIFIER,
-    TS_RS_SCOPED_USE_LIST,
-    TS_RS_SELF_PARAMETER,
-    TS_RS_SOURCE_FILE,
-    TS_RS_STRUCT_EXPRESSION,
-    TS_RS_STRUCT_ITEM,
-    TS_RS_TOKEN_DOT,
-    TS_RS_TRAIT_ITEM,
-    TS_RS_TRY_EXPRESSION,
-    TS_RS_TUPLE_STRUCT_PATTERN,
-    TS_RS_TYPE_ARGUMENTS,
-    TS_RS_TYPE_ITEM,
-    TS_RS_UNION_ITEM,
-    TS_RS_USE_AS_CLAUSE,
-    TS_RS_USE_DECLARATION,
-    TS_RS_USE_LIST,
-    TS_RS_USE_WILDCARD,
-    TS_SATISFIES_EXPRESSION,
-    TS_SCALA_CALL_EXPRESSION,
-    TS_SCALA_CLASS_DEFINITION,
-    TS_SCALA_COMPILATION_UNIT,
-    TS_SCALA_FIELD_EXPRESSION,
-    TS_SCALA_FUNCTION_DECLARATION,
-    TS_SCALA_FUNCTION_DEFINITION,
-    TS_SCALA_GENERIC_FUNCTION,
-    TS_SCALA_IMPORT_DECLARATION,
-    TS_SCALA_INFIX_EXPRESSION,
-    TS_SCALA_OBJECT_DEFINITION,
-    TS_SCALA_TRAIT_DEFINITION,
-    TS_SCOPED_IDENTIFIER,
-    TS_SCOPED_TYPE_IDENTIFIER,
-    TS_SELECTOR_EXPRESSION,
-    TS_SHORTHAND_PROPERTY_IDENTIFIER_PATTERN,
-    TS_SPREAD_PARAMETER,
-    TS_STATEMENT_BLOCK,
-    TS_STATIC,
-    TS_STRING,
-    TS_STRING_LITERAL,
-    TS_STRUCT_SPECIFIER,
-    TS_SUPER,
-    TS_SYSTEM_LIB_STRING,
-    TS_TEMPLATE_ARGUMENT_LIST,
-    TS_TEMPLATE_FUNCTION,
-    TS_TEMPLATE_TYPE,
-    TS_THIS,
-    TS_TRUE,
-    TS_TYPE_ALIAS_DECLARATION,
-    TS_TYPE_DESCRIPTOR,
-    TS_TYPE_IDENTIFIER,
-    TS_TYPE_LIST,
-    TS_TYPE_PARAMETER,
-    TS_UNION_SPECIFIER,
-    TS_USE_DECLARATION,
-    TS_VARIABLE_DECLARATION,
-    TS_VARIABLE_DECLARATOR,
-    TS_VIRTUAL,
-    TS_VOID_TYPE,
-    TS_WILDCARD_IMPORT,
-    TYPE_INFERENCE_BASE_MODEL,
-    TYPE_INFERENCE_LIST,
-    UNIXCODER_BUFFER_BIAS,
-    UNIXCODER_MASK_TOKEN,
-    UNIXCODER_MAX_CONTEXT,
-    UNIXCODER_MODEL,
-    V1_PATH,
-    VERSION_SPLIT_EQ,
-    VERSION_SPLIT_GTE,
-    VERSION_SPLIT_LT,
-    Architecture,
-    AuditCheck,
-    CppFrontend,
-    CppNodeType,
-    DeadCodeFormat,
-    EventType,
-    KeyBinding,
-    MCPEnvVar,
-    MCPParamName,
-    MCPSchemaField,
-    MCPSchemaType,
-    MCPToolName,
-    MCPTransport,
-    PyInstallerPackage,
-    QueryFormat,
-    UniXcoderMode,
-)
-from .graph import (
-    _NODE_LABEL_UNIQUE_KEYS as _NODE_LABEL_UNIQUE_KEYS,
-)
-from .graph import (
-    CYPHER_ALL_DEFINITION_QNS,
-    CYPHER_ALL_FILE_PATHS,
-    CYPHER_ALL_FOLDER_PATHS,
-    CYPHER_ALL_INHERITS,
-    CYPHER_ALL_MODULE_PATHS_INTERNAL,
-    CYPHER_ALLOWED_PROCEDURE_PREFIXES,
-    CYPHER_BACKTICK,
-    CYPHER_DANGEROUS_KEYWORDS,
-    CYPHER_DEFAULT_LIMIT,
-    CYPHER_DELETE_CALLS,
-    CYPHER_DELETE_FILE,
-    CYPHER_DELETE_FOLDER,
-    CYPHER_DELETE_MODULE,
-    CYPHER_DELETE_ORPHAN_EXTERNAL_MODULES,
-    CYPHER_INBOUND_EDGES,
-    CYPHER_MATCH_KEYWORD,
-    CYPHER_MEMORY_LIMIT_SUFFIX,
-    CYPHER_MEMORY_LIMIT_TOKEN,
-    CYPHER_PARAM_PATHS,
-    CYPHER_PREFIX,
-    CYPHER_QUERY_EMBEDDINGS,
-    CYPHER_QUERY_PROJECT_NODE_IDS,
-    CYPHER_SEMICOLON,
-    KEY_ABSOLUTE_PATH,
-    KEY_BASE_INDEX,
-    KEY_BASE_QN,
-    KEY_CALLER_LABEL,
-    KEY_CALLER_QN,
-    KEY_CHILD_QN,
-    KEY_CREATED,
-    KEY_DECORATORS,
-    KEY_DOCSTRING,
-    KEY_END_LINE,
-    KEY_EXPORTED_AT,
-    KEY_EXTENSION,
-    KEY_FROM_ID,
-    KEY_FROM_VAL,
-    KEY_IMPLEMENTS_MODULE,
-    KEY_IS_EXPORTED,
-    KEY_IS_PROPERTY,
-    KEY_LABEL,
-    KEY_LABELS,
-    KEY_METADATA,
-    KEY_MODULE_TYPE,
-    KEY_NAME,
-    KEY_NODE_ID,
-    KEY_NODE_LABELS,
-    KEY_PARAMETERS,
-    KEY_PARSER,
-    KEY_PATH,
-    KEY_PREFIX,
-    KEY_PROJECT_NAME,
-    KEY_PROPERTIES,
-    KEY_PROPS,
-    KEY_QUALIFIED_NAME,
-    KEY_QUERY,
-    KEY_REL,
-    KEY_RELATIONSHIP_TYPES,
-    KEY_RELATIONSHIPS,
-    KEY_RESPONSE,
-    KEY_START_LINE,
-    KEY_TARGET_LABEL,
-    KEY_TARGET_QN,
-    KEY_TO_ID,
-    KEY_TO_VAL,
-    KEY_TOTAL_RELATIONSHIPS,
-    KEY_TYPE,
-    KEY_VERSION_SPEC,
-    NODE_PROJECT,
-    NODE_UNIQUE_CONSTRAINTS,
-    ONEOF_CLASS,
-    ONEOF_ENUM,
-    ONEOF_EXTERNAL_MODULE,
-    ONEOF_EXTERNAL_PACKAGE,
-    ONEOF_FILE,
-    ONEOF_FOLDER,
-    ONEOF_FUNCTION,
-    ONEOF_INTERFACE,
-    ONEOF_METHOD,
-    ONEOF_MODULE,
-    ONEOF_MODULE_IMPLEMENTATION,
-    ONEOF_MODULE_INTERFACE,
-    ONEOF_PACKAGE,
-    ONEOF_PROJECT,
-    ONEOF_TYPE,
-    ONEOF_UNION,
-    PAYLOAD_NODE_ID,
-    PAYLOAD_QUALIFIED_NAME,
-    REL_TYPE_CALLS,
-    TRIE_INTERNAL_PREFIX,
-    TRIE_QN_KEY,
-    TRIE_TYPE_KEY,
-    NodeLabel,
-    RelationshipType,
-    UniqueKeyType,
-)
-from .graph import (
-    _missing_keys as _missing_keys,
-)
-from .languages import (
-    BINARY_EXTENSIONS,
-    C_EXTENSIONS,
-    CPP_EXTENSIONS,
-    CPP_MODULE_EXTENSIONS,
-    EXT_C,
-    EXT_CC,
-    EXT_CCM,
-    EXT_CLASS,
-    EXT_CPP,
-    EXT_CPPM,
-    EXT_CXX,
-    EXT_GO,
-    EXT_H,
-    EXT_HH,
-    EXT_HPP,
-    EXT_HXX,
-    EXT_IXX,
-    EXT_JAVA,
-    EXT_JS,
-    EXT_JSX,
-    EXT_LUA,
-    EXT_PHP,
-    EXT_PY,
-    EXT_RS,
-    EXT_SC,
-    EXT_SCALA,
-    EXT_TS,
-    EXT_TSX,
-    GO_EXTENSIONS,
-    JAVA_EXTENSIONS,
-    JS_EXTENSIONS,
-    JS_TS_MODULE_EXTENSIONS,
-    LANG_PROMPT_EXTENSIONS,
-    LANG_TABLE_COL_EXTENSIONS,
-    LANGUAGE_METADATA,
-    LUA_EXTENSIONS,
-    MIME_TYPE_FALLBACK,
-    MIME_TYPE_PDF,
-    MULTIMODAL_EXTENSIONS,
-    PHP_EXTENSIONS,
-    PY_EXTENSIONS,
-    RS_EXTENSIONS,
-    SCALA_EXTENSIONS,
-    TS_EXTENSIONS,
-    TSX_EXTENSIONS,
-    LanguageMetadata,
-    LanguageStatus,
-    SupportedLanguage,
-)
-from .operators import (
-    CPP_IMPORT_PARTITION_PREFIX,
-    CPP_OPERATOR_PREFIX,
-    CPP_PARTITION_PREFIX,
-    DEFAULT_ROOT_DECORATORS,
-    GO_ROOT_FUNCTION_NAMES,
-    JS_EXTERNAL_IMPORT_SCHEMES,
-    JS_INDEX_STEM,
-    PATH_PARENT_PREFIX,
-    PATH_RELATIVE_PREFIX,
-    PROTOCOL_BASE_QNS,
-    RUST_ROOT_FUNCTION_NAMES,
-    RUST_TRAIT_METHOD_NAMES,
-    TEST_PATH_PATTERNS,
-    TS_ALIAS_SKIP_DIRS,
-    TS_BASE_URL_KEY,
-    TS_COMPILER_OPTIONS_KEY,
-    TS_PATHS_KEY,
-    TSCONFIG_FILENAMES,
-)
-
-__all__ = [
-    "ABSTRACT_DECORATORS",
-    "ALLOWED_COMMENT_MARKERS",
-    "ANTHROPIC_API_VERSION",
-    "ANTHROPIC_COUNT_TIMEOUT_S",
-    "ANTHROPIC_COUNT_TOKENS_URL",
-    "ANTHROPIC_HEADER_API_KEY",
-    "ANTHROPIC_HEADER_VERSION",
-    "ARG_COMMAND",
-    "ARG_CONTENT",
-    "ARG_FILE_PATH",
-    "ARG_REPLACEMENT_CODE",
-    "ARG_TARGET_CODE",
-    "ATTR_TYPE_INFERENCE_IN_PROGRESS",
-    "AUDIT_DETAIL_DANGLING",
-    "AUDIT_DETAIL_MISSING_REQUIRED",
-    "AUDIT_DETAIL_MISSING_REQUIRED_LIVE",
-    "AUDIT_DETAIL_ORPHAN",
-    "AUDIT_DETAIL_ORPHAN_COUNT",
-    "AUDIT_DETAIL_UNDOCUMENTED_LABEL",
-    "AUDIT_DETAIL_UNDOCUMENTED_PROPERTY",
-    "AUDIT_DETAIL_UNDOCUMENTED_PROPERTY_LIVE",
-    "AUDIT_DETAIL_UNDOCUMENTED_RELATIONSHIP",
-    "Architecture",
-    "AuditCheck",
-    "BINARY_EXTENSIONS",
-    "BINARY_FILE_PERMISSION",
-    "BINARY_NAME_TEMPLATE",
-    "BINDINGS_DIR",
-    "BUILD_EXT_CMD",
-    "BUILTIN_PREFIX",
-    "BYTES_PER_MB",
-    "BYTES_PER_MB_FLOAT",
-    "CALL_NODES_BASIC",
-    "CALL_NODES_METHOD",
-    "CALL_NODES_OPERATOR",
-    "CALL_NODES_SPECIAL",
-    "CAPTURE_ARROW_FUNCTION",
-    "CAPTURE_CALL",
-    "CAPTURE_CHILD_CLASS",
-    "CAPTURE_CLASS",
-    "CAPTURE_CONSTRUCTOR_NAME",
-    "CAPTURE_EXPORTS_OBJ",
-    "CAPTURE_EXPORTS_PROP",
-    "CAPTURE_EXPORT_FUNCTION",
-    "CAPTURE_EXPORT_NAME",
-    "CAPTURE_FUNC",
-    "CAPTURE_FUNCTION",
-    "CAPTURE_FUNCTION_EXPR",
-    "CAPTURE_IMPORT",
-    "CAPTURE_IMPORT_FROM",
-    "CAPTURE_MEMBER_EXPR",
-    "CAPTURE_METHOD_FUNCTION",
-    "CAPTURE_METHOD_NAME",
-    "CAPTURE_MODULE_OBJ",
-    "CAPTURE_PARENT_CLASS",
-    "CAPTURE_PROTOTYPE_KEYWORD",
-    "CAPTURE_VARIABLE_DECLARATOR",
-    "CHAR_ANGLE_CLOSE",
-    "CHAR_ANGLE_OPEN",
-    "CHAR_COLON",
-    "CHAR_COMMA",
-    "CHAR_HYPHEN",
-    "CHAR_PAREN_CLOSE",
-    "CHAR_PAREN_OPEN",
-    "CHAR_SEMICOLON",
-    "CHAR_SPACE",
-    "CHAR_UNDERSCORE",
-    "CLASS_NODES_BASIC",
-    "CLASS_NODES_ENUM",
-    "CLASS_NODES_INTERFACE",
-    "CLASS_NODES_STRUCT",
-    "CLASS_NODES_TYPE_ALIAS",
-    "CLASS_NODES_UNION",
-    "CLI_DEADCODE_COL_KIND",
-    "CLI_DEADCODE_COL_LINES",
-    "CLI_DEADCODE_COL_QUALIFIED_NAME",
-    "CLI_DEADCODE_CONNECTING",
-    "CLI_DEADCODE_LINE_RANGE",
-    "CLI_DEADCODE_NONE",
-    "CLI_DEADCODE_SUMMARY",
-    "CLI_DEADCODE_TABLE_TITLE",
-    "CLI_DEADCODE_WRITTEN",
-    "CLI_ERR_CONFIG",
-    "CLI_ERR_DEADCODE_AMBIGUOUS_PROJECT",
-    "CLI_ERR_DEADCODE_FAILED",
-    "CLI_ERR_DEADCODE_NO_PROJECTS",
-    "CLI_ERR_DELETE_PROJECT_FAILED",
-    "CLI_ERR_EXPORT_FAILED",
-    "CLI_ERR_INDEXING",
-    "CLI_ERR_JSON_REQUIRES_ASK_AGENT",
-    "CLI_ERR_LOAD_GRAPH",
-    "CLI_ERR_MCP_SERVER",
-    "CLI_ERR_ONLY_JSON",
-    "CLI_ERR_OUTPUT_REQUIRES_UPDATE",
-    "CLI_ERR_PATH_NOT_DIR",
-    "CLI_ERR_PATH_NOT_EXISTS",
-    "CLI_ERR_PROJECT_NAME_REQUIRED",
-    "CLI_ERR_PROJECT_NOT_FOUND",
-    "CLI_ERR_STARTUP",
-    "CLI_ERR_STATS_FAILED",
-    "CLI_MSG_APP_TERMINATED",
-    "CLI_MSG_AUTO_EXCLUDE",
-    "CLI_MSG_CLEANING_DB",
-    "CLI_MSG_CLEANING_HASH_CACHE",
-    "CLI_MSG_CLEAN_DONE",
-    "CLI_MSG_CONNECTING_MEMGRAPH",
-    "CLI_MSG_CONNECTING_STATS",
-    "CLI_MSG_DELETING_PROJECT",
-    "CLI_MSG_EXPORTING_DATA",
-    "CLI_MSG_EXPORTING_TO",
-    "CLI_MSG_GRAPH_SUMMARY",
-    "CLI_MSG_GRAPH_UPDATED",
-    "CLI_MSG_HINT_TARGET_REPO",
-    "CLI_MSG_INDEXING_AT",
-    "CLI_MSG_INDEXING_DONE",
-    "CLI_MSG_MCP_TERMINATED",
-    "CLI_MSG_OPTIMIZATION_TERMINATED",
-    "CLI_MSG_OUTPUT_TO",
-    "CLI_MSG_PROJECT_DELETED",
-    "CLI_MSG_SYNCING_GRAPH",
-    "CLI_MSG_SYNC_DONE",
-    "CLI_MSG_SYNC_SKIPPED",
-    "CLI_MSG_UPDATING_GRAPH",
-    "CLI_MSG_VERSION",
-    "CLI_MSG_WORKSPACE_EMPTY",
-    "CLI_MSG_WORKSPACE_SYNCING",
-    "CLI_MSG_WORKSPACE_SYNC_REPO",
-    "CLI_STATS_COL_COUNT",
-    "CLI_STATS_COL_NODE_TYPE",
-    "CLI_STATS_COL_REL_TYPE",
-    "CLI_STATS_NODE_TITLE",
-    "CLI_STATS_REL_TITLE",
-    "CLI_STATS_TOTAL_NODES",
-    "CLI_STATS_TOTAL_RELS",
-    "CLI_STATS_UNKNOWN",
-    "CLI_WARN_NOT_GIT_REPO",
-    "COMMENT_CHAR",
-    "CONFIRM_DISABLED",
-    "CONFIRM_ENABLED",
-    "CONTENT_TYPE_JSON",
-    "CPP_CLASS_TYPES",
-    "CPP_COMPOUND_TYPES",
-    "CPP_DECLARATOR_SUFFIX",
-    "CPP_DESTRUCTOR_PREFIX",
-    "CPP_EXPORTED_CLASS_KEYWORDS",
-    "CPP_EXPORT_CLASS_PREFIX",
-    "CPP_EXPORT_MODULE_PREFIX",
-    "CPP_EXPORT_PREFIXES",
-    "CPP_EXPORT_STRUCT_PREFIX",
-    "CPP_EXPORT_TEMPLATE_PREFIX",
-    "CPP_EXPORT_UNION_PREFIX",
-    "CPP_EXTENSIONS",
-    "CPP_FALLBACK_DESTRUCTOR",
-    "CPP_FALLBACK_OPERATOR",
-    "CPP_IMPL_SUFFIX",
-    "CPP_IMPORT_NODES",
-    "CPP_IMPORT_PARTITION_PREFIX",
-    "CPP_KEYWORD_CLASS",
-    "CPP_KEYWORD_STRUCT",
-    "CPP_MODULE_EXTENSIONS",
-    "CPP_MODULE_PATH_MARKERS",
-    "CPP_MODULE_PREFIX",
-    "CPP_MODULE_PRIVATE_PREFIX",
-    "CPP_MODULE_TYPE_IMPLEMENTATION",
-    "CPP_MODULE_TYPE_INTERFACE",
-    "CPP_NAME_NODE_TYPES",
-    "CPP_NESTED_SCOPE_NODE_TYPES",
-    "CPP_OPERATOR_PREFIX",
-    "CPP_OPERATOR_SYMBOL_MAP",
-    "CPP_OPERATOR_TEXT_PREFIX",
-    "CPP_PARTITION_PREFIX",
-    "CPP_PREFIX_HAS",
-    "CPP_PREFIX_IS",
-    "CPP_STDLIB_ENTITIES",
-    "CPP_STD_NAMESPACE",
-    "CPP_STD_PREFIX",
-    "CPP_TYPE_PARAMETER_DECL_TYPES",
-    "CPP_TYPE_SPECIFIER_NODE_TYPES",
-    "CSPROJ_SUFFIX",
-    "CYPHER_ALLOWED_PROCEDURE_PREFIXES",
-    "CYPHER_ALL_DEFINITION_QNS",
-    "CYPHER_ALL_FILE_PATHS",
-    "CYPHER_ALL_FOLDER_PATHS",
-    "CYPHER_ALL_INHERITS",
-    "CYPHER_ALL_MODULE_PATHS_INTERNAL",
-    "CYPHER_ALL_MODULE_QNS",
-    "CYPHER_BACKTICK",
-    "CYPHER_DANGEROUS_KEYWORDS",
-    "CYPHER_DEFAULT_LIMIT",
-    "CYPHER_DELETE_CALLS",
-    "CYPHER_DELETE_FILE",
-    "CYPHER_DELETE_FOLDER",
-    "CYPHER_DELETE_MODULE",
-    "CYPHER_DELETE_ORPHAN_EXTERNAL_MODULES",
-    "CYPHER_INBOUND_EDGES",
-    "CYPHER_MATCH_KEYWORD",
-    "CYPHER_MEMORY_LIMIT_SUFFIX",
-    "CYPHER_MEMORY_LIMIT_TOKEN",
-    "CYPHER_PARAM_PATHS",
-    "CYPHER_PREFIX",
-    "CYPHER_QUERY_EMBEDDINGS",
-    "CYPHER_QUERY_PROJECT_NODE_IDS",
-    "CYPHER_SEMICOLON",
-    "C_EXTENSIONS",
-    "C_NAME_NODE_TYPES",
-    "Color",
-    "CppFrontend",
-    "CppNodeType",
-    "DECORATOR_AT",
-    "DEFAULT_API_KEY",
-    "DEFAULT_CONTEXT_WINDOW",
-    "DEFAULT_DEBOUNCE_SECONDS",
-    "DEFAULT_MAX_WAIT_SECONDS",
-    "DEFAULT_MODEL",
-    "DEFAULT_MODEL_ROLE",
-    "DEFAULT_NAME",
-    "DEFAULT_REGION",
-    "DEFAULT_ROOT_DECORATORS",
-    "DEFAULT_TABLE_TITLE",
-    "DELIMITER_TOKENS",
-    "DEPENDENCY_FILES",
-    "DEP_ATTR_INCLUDE",
-    "DEP_ATTR_VERSION",
-    "DEP_EXCLUDE_PHP",
-    "DEP_EXCLUDE_PYTHON",
-    "DEP_FILE_CARGO",
-    "DEP_FILE_COMPOSER",
-    "DEP_FILE_GEMFILE",
-    "DEP_FILE_GOMOD",
-    "DEP_FILE_PACKAGE_JSON",
-    "DEP_FILE_PYPROJECT",
-    "DEP_FILE_REQUIREMENTS",
-    "DEP_KEY_DEPENDENCIES",
-    "DEP_KEY_DEV_DEPENDENCIES",
-    "DEP_KEY_DEV_DEPS_JSON",
-    "DEP_KEY_GROUP",
-    "DEP_KEY_OPTIONAL_DEPS",
-    "DEP_KEY_PEER_DEPS",
-    "DEP_KEY_POETRY",
-    "DEP_KEY_PROJECT",
-    "DEP_KEY_REQUIRE",
-    "DEP_KEY_REQUIRE_DEV",
-    "DEP_KEY_TOOL",
-    "DEP_KEY_VERSION",
-    "DEP_XML_PACKAGE_REF",
-    "DICT_KEY_RESULTS",
-    "DIFF_CONTINUATION_PREFIXES",
-    "DIFF_FALLBACK_PATH",
-    "DIFF_GIT_HEADER",
-    "DIFF_LABEL_AFTER",
-    "DIFF_LABEL_BEFORE",
-    "DIR_MTIMES_FILENAME",
-    "DIST_DIR",
-    "DOCSTRING_STRIP_CHARS",
-    "DOC_PROMPT_PREFIX",
-    "DUP_QN_MARKER",
-    "DeadCodeFormat",
-    "DiffMarker",
-    "EMBEDDING_CACHE_FILENAME",
-    "EMBEDDING_DEFAULT_BATCH_SIZE",
-    "EMPTY_PARENS",
-    "ENCODING_UTF8",
-    "ENTITY_CLASS",
-    "ENTITY_CONSTRUCTOR",
-    "ENTITY_FUNCTION",
-    "ENTITY_METHOD",
-    "ENV_ANTHROPIC_API_KEY",
-    "ENV_AZURE_API_KEY",
-    "ENV_AZURE_API_VERSION",
-    "ENV_AZURE_ENDPOINT",
-    "ENV_GOOGLE_API_KEY",
-    "ENV_OPENAI_API_KEY",
-    "ERR_SUBSTR_ALREADY_EXISTS",
-    "ERR_SUBSTR_CONSTRAINT",
-    "ESCAPE_CHAR",
-    "EXCLUDED_DEPENDENCY_NAMES",
-    "EXIT_COMMANDS",
-    "EXT_C",
-    "EXT_CC",
-    "EXT_CCM",
-    "EXT_CLASS",
-    "EXT_CPP",
-    "EXT_CPPM",
-    "EXT_CXX",
-    "EXT_GO",
-    "EXT_H",
-    "EXT_HH",
-    "EXT_HPP",
-    "EXT_HXX",
-    "EXT_IXX",
-    "EXT_JAVA",
-    "EXT_JS",
-    "EXT_JSX",
-    "EXT_LUA",
-    "EXT_PHP",
-    "EXT_PY",
-    "EXT_RS",
-    "EXT_SC",
-    "EXT_SCALA",
-    "EXT_TS",
-    "EXT_TSX",
-    "EventType",
-    "FIELD_ALIAS",
-    "FIELD_API_KEY",
-    "FIELD_ARGUMENTS",
-    "FIELD_BODY",
-    "FIELD_BOUNDS",
-    "FIELD_CONSTRUCTOR",
-    "FIELD_DECLARATOR",
-    "FIELD_DEFINITION",
-    "FIELD_ENDPOINT",
-    "FIELD_FIELD",
-    "FIELD_FUNCTION",
-    "FIELD_INTERFACES",
-    "FIELD_KEY",
-    "FIELD_LEFT",
-    "FIELD_MACRO",
-    "FIELD_MODEL_ID",
-    "FIELD_MODULE_NAME",
-    "FIELD_NAME",
-    "FIELD_OBJECT",
-    "FIELD_OPERAND",
-    "FIELD_OPERATOR",
-    "FIELD_PARAMETERS",
-    "FIELD_PROPERTY",
-    "FIELD_PROVIDER",
-    "FIELD_RECEIVER",
-    "FIELD_RESULT",
-    "FIELD_RETURN_TYPE",
-    "FIELD_RIGHT",
-    "FIELD_SCOPE",
-    "FIELD_SUPERCLASS",
-    "FIELD_SUPERCLASSES",
-    "FIELD_TRAIT",
-    "FIELD_TYPE",
-    "FIELD_VALUE",
-    "FQN_CPP_FUNCTION_TYPES",
-    "FQN_CPP_SCOPE_TYPES",
-    "FQN_C_FUNCTION_TYPES",
-    "FQN_C_SCOPE_TYPES",
-    "FQN_GO_FUNCTION_TYPES",
-    "FQN_GO_SCOPE_TYPES",
-    "FQN_JAVA_FUNCTION_TYPES",
-    "FQN_JAVA_SCOPE_TYPES",
-    "FQN_JS_FUNCTION_TYPES",
-    "FQN_JS_SCOPE_TYPES",
-    "FQN_LUA_FUNCTION_TYPES",
-    "FQN_LUA_SCOPE_TYPES",
-    "FQN_PHP_FUNCTION_TYPES",
-    "FQN_PHP_SCOPE_TYPES",
-    "FQN_PY_FUNCTION_TYPES",
-    "FQN_PY_SCOPE_TYPES",
-    "FQN_RS_FUNCTION_TYPES",
-    "FQN_RS_SCOPE_TYPES",
-    "FQN_SCALA_FUNCTION_TYPES",
-    "FQN_SCALA_SCOPE_TYPES",
-    "FQN_TS_FUNCTION_TYPES",
-    "FQN_TS_SCOPE_TYPES",
-    "FUNCTION_NODES_BASIC",
-    "FUNCTION_NODES_GENERATOR",
-    "FUNCTION_NODES_LAMBDA",
-    "FUNCTION_NODES_METHOD",
-    "FUNCTION_NODES_TEMPLATE",
-    "FileAction",
-    "GEMFILE_GEM_PREFIX",
-    "GITWILDMATCH_STYLE",
-    "GIT_DIR_NAME",
-    "GLOB_ALL",
-    "GOMOD_BLOCK_END",
-    "GOMOD_COMMENT_PREFIX",
-    "GOMOD_REQUIRE_BLOCK_START",
-    "GOMOD_REQUIRE_LINE_PREFIX",
-    "GOOGLE_CLOUD_SCOPE",
-    "GO_EXTENSIONS",
-    "GO_KEYWORD_MODULE",
-    "GO_MOD_COMMENT_PREFIX",
-    "GO_ROOT_FUNCTION_NAMES",
-    "GO_TEST_FILE_SUFFIX",
-    "GRAMMARS_DIR",
-    "GREP_SUGGESTION",
-    "GUARD_INHERITED_METHOD",
-    "GoogleProviderType",
-    "HASH_CACHE_FILENAME",
-    "HEADER_CONTENT_TYPE",
-    "HEALTH_CHECK_API_KEY_CONFIGURED",
-    "HEALTH_CHECK_API_KEY_MISSING_MSG",
-    "HEALTH_CHECK_API_KEY_NOT_CONFIGURED",
-    "HEALTH_CHECK_API_KEY_NOT_SET",
-    "HEALTH_CHECK_API_KEY_SET",
-    "HEALTH_CHECK_DOCKER_EXIT_CODE",
-    "HEALTH_CHECK_DOCKER_FAILED_MSG",
-    "HEALTH_CHECK_DOCKER_NOT_INSTALLED_MSG",
-    "HEALTH_CHECK_DOCKER_NOT_IN_PATH",
-    "HEALTH_CHECK_DOCKER_NOT_RESPONDING_MSG",
-    "HEALTH_CHECK_DOCKER_NOT_RUNNING",
-    "HEALTH_CHECK_DOCKER_RUNNING",
-    "HEALTH_CHECK_DOCKER_RUNNING_MSG",
-    "HEALTH_CHECK_DOCKER_TIMEOUT_ERROR",
-    "HEALTH_CHECK_DOCKER_TIMEOUT_MSG",
-    "HEALTH_CHECK_EXTERNAL_TOOLS",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_ERROR_MSG",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_FAILED",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_OK",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_OK_MSG",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_SEPARATOR",
-    "HEALTH_CHECK_GRAPH_INTEGRITY_VIOLATIONS_MSG",
-    "HEALTH_CHECK_MEMGRAPH_CONNECTED_MSG",
-    "HEALTH_CHECK_MEMGRAPH_CONNECTION_FAILED_MSG",
-    "HEALTH_CHECK_MEMGRAPH_ERROR",
-    "HEALTH_CHECK_MEMGRAPH_FAILED",
-    "HEALTH_CHECK_MEMGRAPH_QUERY",
-    "HEALTH_CHECK_MEMGRAPH_SUCCESSFUL",
-    "HEALTH_CHECK_MEMGRAPH_UNEXPECTED_FAILURE_MSG",
-    "HEALTH_CHECK_TOOLS",
-    "HEALTH_CHECK_TOOL_FAILED_MSG",
-    "HEALTH_CHECK_TOOL_INSTALLED",
-    "HEALTH_CHECK_TOOL_INSTALLED_MSG",
-    "HEALTH_CHECK_TOOL_NOT_INSTALLED",
-    "HEALTH_CHECK_TOOL_NOT_IN_PATH_MSG",
-    "HEALTH_CHECK_TOOL_TIMEOUT_ERROR",
-    "HEALTH_CHECK_TOOL_TIMEOUT_MSG",
-    "HELP_ARG",
-    "HELP_COMMAND",
-    "HIGHER_ORDER_BUILTINS",
-    "HORIZONTAL_SEPARATOR",
-    "HTTP_OK",
-    "IGNORE_PATTERNS",
-    "IGNORE_SUFFIXES",
-    "IIFE_ARROW_PREFIX",
-    "IIFE_FUNC_PREFIX",
-    "IMPORT_CACHE_DIR",
-    "IMPORT_CACHE_FILE",
-    "IMPORT_CACHE_KEY",
-    "IMPORT_CACHE_TTL",
-    "IMPORT_DEFAULT_SUFFIX",
-    "IMPORT_IMPORT",
-    "IMPORT_MODULE_LABEL",
-    "IMPORT_NODES_FROM",
-    "IMPORT_NODES_INCLUDE",
-    "IMPORT_NODES_MODULE",
-    "IMPORT_NODES_STANDARD",
-    "IMPORT_PCALL",
-    "IMPORT_QUALIFIED_NAME",
-    "IMPORT_RELATIONSHIP",
-    "IMPORT_REQUIRE",
-    "IMPORT_STD_PREFIX",
-    "IMPORT_TIMESTAMPS_KEY",
-    "INDEX_INDEX",
-    "INDEX_INIT",
-    "INDEX_LUA_INIT",
-    "INDEX_MOD",
-    "INIT_PY",
-    "INLINE_MODULE_PATH_PREFIX",
-    "INPLACE_FLAG",
-    "INTERACTIVE_BFS_MAX_DEPTH",
-    "INTERACTIVE_COL_NESTED",
-    "INTERACTIVE_COL_NUM",
-    "INTERACTIVE_COL_PATH",
-    "INTERACTIVE_COL_PATTERN",
-    "INTERACTIVE_DEFAULT_GROUP",
-    "INTERACTIVE_EXPAND_SUFFIX",
-    "INTERACTIVE_INSTRUCTIONS_GROUPED",
-    "INTERACTIVE_INSTRUCTIONS_NESTED",
-    "INTERACTIVE_KEEP_ALL",
-    "INTERACTIVE_KEEP_NONE",
-    "INTERACTIVE_NESTED_PLURAL",
-    "INTERACTIVE_NESTED_SINGULAR",
-    "INTERACTIVE_PROMPT_KEEP",
-    "INTERACTIVE_STATUS_CGRIGNORE",
-    "INTERACTIVE_STATUS_CLI",
-    "INTERACTIVE_STATUS_DETECTED",
-    "INTERACTIVE_STYLE_DIM",
-    "INTERACTIVE_TITLE_GROUPED",
-    "INTERACTIVE_TITLE_NESTED",
-    "JAVA_ARRAY_SUFFIX",
-    "JAVA_CALLABLE_ENTITY_TYPES",
-    "JAVA_CLASS_MODIFIERS",
-    "JAVA_CLASS_NODE_TYPES",
-    "JAVA_CREATE_PATTERN",
-    "JAVA_DECLARATION_SUFFIX",
-    "JAVA_EXTENSIONS",
-    "JAVA_FIELD_MODIFIERS",
-    "JAVA_GETTER_PATTERN",
-    "JAVA_HAS_PATTERN",
-    "JAVA_HEURISTIC_ORDER",
-    "JAVA_HEURISTIC_USER",
-    "JAVA_ID_PATTERN",
-    "JAVA_IS_PATTERN",
-    "JAVA_JVM_LANGUAGES",
-    "JAVA_KEYWORD_SUPER",
-    "JAVA_KEYWORD_THIS",
-    "JAVA_LANG_CLASS_NAMES",
-    "JAVA_LANG_PREFIX",
-    "JAVA_LENGTH_PATTERN",
-    "JAVA_MAIN_METHOD_NAME",
-    "JAVA_MAIN_PARAM_ARRAY",
-    "JAVA_MAIN_PARAM_TYPE",
-    "JAVA_MAIN_PARAM_VARARGS",
-    "JAVA_METHOD_MODIFIERS",
-    "JAVA_METHOD_NODE_TYPES",
-    "JAVA_MODIFIER_ABSTRACT",
-    "JAVA_MODIFIER_FINAL",
-    "JAVA_MODIFIER_PRIVATE",
-    "JAVA_MODIFIER_PROTECTED",
-    "JAVA_MODIFIER_PUBLIC",
-    "JAVA_MODIFIER_STATIC",
-    "JAVA_MODIFIER_SYNCHRONIZED",
-    "JAVA_MODIFIER_TRANSIENT",
-    "JAVA_MODIFIER_VOLATILE",
-    "JAVA_NAME_PATTERN",
-    "JAVA_NEW_PATTERN",
-    "JAVA_ORDER_PATTERN",
-    "JAVA_PATH_JAVA",
-    "JAVA_PATH_KOTLIN",
-    "JAVA_PATH_MAIN",
-    "JAVA_PATH_SCALA",
-    "JAVA_PATH_SRC",
-    "JAVA_PATH_TEST",
-    "JAVA_PRIMITIVE_TYPES",
-    "JAVA_SERIALIZATION_METHOD_NAMES",
-    "JAVA_SIZE_PATTERN",
-    "JAVA_SRC_FOLDERS",
-    "JAVA_STDLIB_CLASSES",
-    "JAVA_STDLIB_PREFIXES",
-    "JAVA_SUFFIX_BUILDER",
-    "JAVA_SUFFIX_ERROR",
-    "JAVA_SUFFIX_EXCEPTION",
-    "JAVA_SUFFIX_INTERFACE",
-    "JAVA_TYPE_BOOLEAN",
-    "JAVA_TYPE_CONSTRUCTOR",
-    "JAVA_TYPE_DOUBLE",
-    "JAVA_TYPE_INT",
-    "JAVA_TYPE_LONG",
-    "JAVA_TYPE_METHOD",
-    "JAVA_TYPE_OBJECT",
-    "JAVA_TYPE_STRING",
-    "JAVA_TYPE_STRING_FQN",
-    "JAVA_USER_PATTERN",
-    "JAVA_VISIBILITY_PACKAGE",
-    "JAVA_VISIBILITY_PRIVATE",
-    "JAVA_VISIBILITY_PROTECTED",
-    "JAVA_VISIBILITY_PUBLIC",
-    "JAVA_WRAPPER_TYPES",
-    "JSON_EMPTY_OBJECT",
-    "JSON_INDENT",
-    "JSON_KEY_ENTITY_TYPE",
-    "JSON_KEY_HAS_ENTITY",
-    "JS_ARROW_NAME_CLIMB_BOUNDARY",
-    "JS_ASSIGNMENT_ARROW_QUERY",
-    "JS_ASSIGNMENT_FUNCTION_QUERY",
-    "JS_BUILTIN_PATTERNS",
-    "JS_BUILTIN_TYPES",
-    "JS_CALL_RESULT_VALUE_TYPES",
-    "JS_COMMONJS_DESTRUCTURE_QUERY",
-    "JS_COMMONJS_EXPORTS_FUNCTION_QUERY",
-    "JS_COMMONJS_MODULE_EXPORTS_QUERY",
-    "JS_CREATE_METHOD",
-    "JS_ES6_EXPORT_CONST_QUERY",
-    "JS_ES6_EXPORT_FUNCTION_QUERY",
-    "JS_EXPORTS_KEYWORD",
-    "JS_EXPORT_TYPE_COMMONJS",
-    "JS_EXPORT_TYPE_COMMONJS_MODULE",
-    "JS_EXPORT_TYPE_ES6_FUNCTION",
-    "JS_EXPORT_TYPE_ES6_FUNCTION_DECL",
-    "JS_EXTENSIONS",
-    "JS_EXTERNAL_IMPORT_SCHEMES",
-    "JS_FUNCTION_PROTOTYPE_METHODS",
-    "JS_FUNCTION_PROTOTYPE_SUFFIXES",
-    "JS_GLOBAL_CLASS_NAMES",
-    "JS_INDEX_STEM",
-    "JS_LOCALS_PATTERN",
-    "JS_METHOD_APPLY",
-    "JS_METHOD_BIND",
-    "JS_METHOD_CALL",
-    "JS_METHOD_DEF_QUERY",
-    "JS_MODULE_KEYWORD",
-    "JS_MODULE_RECEIVERS",
-    "JS_NAME_NODE_TYPES",
-    "JS_OBJECT_ARROW_QUERY",
-    "JS_OBJECT_METHOD_QUERY",
-    "JS_OBJECT_NAME",
-    "JS_PROTOTYPE_INHERITANCE_QUERY",
-    "JS_PROTOTYPE_KEYWORD",
-    "JS_PROTOTYPE_METHOD_QUERY",
-    "JS_REQUIRE_KEYWORD",
-    "JS_SUFFIX_APPLY",
-    "JS_SUFFIX_BIND",
-    "JS_SUFFIX_CALL",
-    "JS_TS_ALL_EXTENSIONS",
-    "JS_TS_CLASS_NODES",
-    "JS_TS_FUNCTION_NODES",
-    "JS_TS_IMPORT_NODES",
-    "JS_TS_LANGUAGES",
-    "JS_TS_MODULE_EXTENSIONS",
-    "JS_TS_PARENT_REF_TYPES",
-    "KEYWORD_CONSTRUCTOR",
-    "KEYWORD_SELF",
-    "KEYWORD_SUPER",
-    "KEY_ABSOLUTE_PATH",
-    "KEY_BASE_INDEX",
-    "KEY_BASE_QN",
-    "KEY_CALLER_LABEL",
-    "KEY_CALLER_QN",
-    "KEY_CHILD_QN",
-    "KEY_CREATED",
-    "KEY_DECORATORS",
-    "KEY_DOCSTRING",
-    "KEY_END_LINE",
-    "KEY_EXPORTED_AT",
-    "KEY_EXTENSION",
-    "KEY_FROM_ID",
-    "KEY_FROM_VAL",
-    "KEY_IMPLEMENTS_MODULE",
-    "KEY_IS_EXPORTED",
-    "KEY_IS_MACRO",
-    "KEY_IS_PROPERTY",
-    "KEY_LABEL",
-    "KEY_LABELS",
-    "KEY_METADATA",
-    "KEY_MODULE_TYPE",
-    "KEY_NAME",
-    "KEY_NODES",
-    "KEY_NODE_ID",
-    "KEY_NODE_LABELS",
-    "KEY_OVERRIDES_EXTERNAL",
-    "KEY_PARAMETERS",
-    "KEY_PARSER",
-    "KEY_PATH",
-    "KEY_PREFIX",
-    "KEY_PROJECT_NAME",
-    "KEY_PROPERTIES",
-    "KEY_PROPS",
-    "KEY_QUALIFIED_NAME",
-    "KEY_QUERY",
-    "KEY_REL",
-    "KEY_RELATIONSHIPS",
-    "KEY_RELATIONSHIP_TYPES",
-    "KEY_RESPONSE",
-    "KEY_START_LINE",
-    "KEY_TARGET_LABEL",
-    "KEY_TARGET_QN",
-    "KEY_TOTAL_NODES",
-    "KEY_TOTAL_RELATIONSHIPS",
-    "KEY_TO_ID",
-    "KEY_TO_VAL",
-    "KEY_TYPE",
-    "KEY_VERSION_SPEC",
-    "KeyBinding",
-    "LANGUAGE_METADATA",
-    "LANG_ATTR_PHP",
-    "LANG_ATTR_PREFIX",
-    "LANG_ATTR_TSX",
-    "LANG_ATTR_TYPESCRIPT",
-    "LANG_CALL_KEYWORDS",
-    "LANG_CALL_KEYWORD_EXCLUDE",
-    "LANG_CLASS_KEYWORDS",
-    "LANG_CONFIG_FILE",
-    "LANG_DEFAULT_CALL_NODES",
-    "LANG_DEFAULT_CLASS_NODES",
-    "LANG_DEFAULT_FUNCTION_NODES",
-    "LANG_DEFAULT_GRAMMAR_URL",
-    "LANG_DEFAULT_MODULE_NODES",
-    "LANG_ELLIPSIS",
-    "LANG_ERR_CONFIG_NOT_FOUND",
-    "LANG_ERR_CUSTOM_URL_HINT",
-    "LANG_ERR_GIT",
-    "LANG_ERR_MANUAL_REMOVE_HINT",
-    "LANG_ERR_MISSING_ARGS",
-    "LANG_ERR_NODE_TYPES_WARNING",
-    "LANG_ERR_NO_GRAMMARS_WARNING",
-    "LANG_ERR_PARSE_NODE_TYPES",
-    "LANG_ERR_REINSTALL_FAILED",
-    "LANG_ERR_REMOVE_CONFIG",
-    "LANG_ERR_REMOVE_SUBMODULE",
-    "LANG_ERR_REPO_NOT_FOUND",
-    "LANG_ERR_TREE_SITTER_JSON_WARNING",
-    "LANG_ERR_UPDATE_CONFIG",
-    "LANG_EXCLUSION_KEYWORDS",
-    "LANG_FALLBACK_MANUAL_ADD",
-    "LANG_FALLBACK_METHOD_NODE",
-    "LANG_FUNCTION_KEYWORDS",
-    "LANG_GITMODULES_FILE",
-    "LANG_GITMODULES_REGEX",
-    "LANG_GIT_MODULES_PATH",
-    "LANG_GIT_SUFFIX",
-    "LANG_GRAMMARS_DIR",
-    "LANG_MODULE_KEYWORDS",
-    "LANG_MSG_ADDING_SUBMODULE",
-    "LANG_MSG_AUTO_DETECTED_EXT",
-    "LANG_MSG_AUTO_DETECTED_LANG",
-    "LANG_MSG_AVAILABLE_LANGS",
-    "LANG_MSG_AVAILABLE_NODES",
-    "LANG_MSG_CALLS",
-    "LANG_MSG_CATEGORY_FORMAT",
-    "LANG_MSG_CLASSES",
-    "LANG_MSG_CLEANED_MODULES",
-    "LANG_MSG_CLEANUP_CANCELLED",
-    "LANG_MSG_CLEANUP_COMPLETE",
-    "LANG_MSG_COMMON_ISSUES",
-    "LANG_MSG_CUSTOM_URL_WARNING",
-    "LANG_MSG_EDIT_HINT",
-    "LANG_MSG_FOUND_NODE_TYPES",
-    "LANG_MSG_FOUND_ORPHANS",
-    "LANG_MSG_FUNCTIONS",
-    "LANG_MSG_ISSUE_CALL_TYPES",
-    "LANG_MSG_ISSUE_CLASS_TYPES",
-    "LANG_MSG_ISSUE_MISCLASSIFIED",
-    "LANG_MSG_ISSUE_MISSING",
-    "LANG_MSG_KEEPING_SUBMODULE",
-    "LANG_MSG_LANG_ADDED",
-    "LANG_MSG_LANG_NOT_FOUND",
-    "LANG_MSG_LANG_REMOVED",
-    "LANG_MSG_LIST_HINT",
-    "LANG_MSG_MAPPED_CATEGORIES",
-    "LANG_MSG_MODULES",
-    "LANG_MSG_NO_GITMODULES",
-    "LANG_MSG_NO_MODULES_DIR",
-    "LANG_MSG_NO_ORPHANS",
-    "LANG_MSG_NO_SUBMODULE",
-    "LANG_MSG_READDING_SUBMODULE",
-    "LANG_MSG_REINSTALL_SUCCESS",
-    "LANG_MSG_REMOVED_FROM_CONFIG",
-    "LANG_MSG_REMOVED_ORPHAN",
-    "LANG_MSG_REMOVING_ENTRY",
-    "LANG_MSG_REMOVING_SUBMODULE",
-    "LANG_MSG_REVIEW_HINT",
-    "LANG_MSG_REVIEW_PROMPT",
-    "LANG_MSG_SEMANTIC_CATEGORIES",
-    "LANG_MSG_SUBMODULE_EXISTS",
-    "LANG_MSG_SUBMODULE_REMOVED",
-    "LANG_MSG_SUBMODULE_SUCCESS",
-    "LANG_MSG_UPDATED_CONFIG",
-    "LANG_MSG_USING_DEFAULT_URL",
-    "LANG_MSG_USING_LANG_NAME",
-    "LANG_NODE_TYPES_JSON",
-    "LANG_PROMPT_CALLS",
-    "LANG_PROMPT_CLASSES",
-    "LANG_PROMPT_COMMON_NAME",
-    "LANG_PROMPT_CONTINUE",
-    "LANG_PROMPT_EXTENSIONS",
-    "LANG_PROMPT_FUNCTIONS",
-    "LANG_PROMPT_LANGUAGE_NAME",
-    "LANG_PROMPT_MODULES",
-    "LANG_PROMPT_REMOVE_ORPHANS",
-    "LANG_SRC_DIR",
-    "LANG_TABLE_COL_CALL_TYPES",
-    "LANG_TABLE_COL_CLASS_TYPES",
-    "LANG_TABLE_COL_EXTENSIONS",
-    "LANG_TABLE_COL_FUNCTION_TYPES",
-    "LANG_TABLE_COL_LANGUAGE",
-    "LANG_TABLE_PLACEHOLDER",
-    "LANG_TABLE_TITLE",
-    "LANG_TREE_SITTER_JSON",
-    "LANG_TREE_SITTER_URL_MARKER",
-    "LOG_FORMAT",
-    "LOG_LEVEL_ERROR",
-    "LOG_LEVEL_INFO",
-    "LUA_DEFAULT_VAR_TYPES",
-    "LUA_EXTENSIONS",
-    "LUA_METHOD_SEPARATOR",
-    "LUA_STATEMENT_SUFFIX",
-    "LUA_STDLIB_MODULES",
-    "LUA_STRING_TYPES",
-    "LanguageMetadata",
-    "LanguageStatus",
-    "MARKDOWN_FENCE",
-    "MARKDOWN_FENCE_DIFF",
-    "MCPEnvVar",
-    "MCPParamName",
-    "MCPSchemaField",
-    "MCPSchemaType",
-    "MCPToolName",
-    "MCPTransport",
-    "MCP_ASK_AGENT_ERROR",
-    "MCP_CONTENT_TYPE_TEXT",
-    "MCP_DEFAULT_DIRECTORY",
-    "MCP_INDEX_ERROR",
-    "MCP_INDEX_SUCCESS",
-    "MCP_INDEX_SUCCESS_PROJECT",
-    "MCP_JSON_INDENT",
-    "MCP_KEY_ERROR",
-    "MCP_KEY_ERROR_MESSAGE",
-    "MCP_KEY_FOUND",
-    "MCP_KEY_QUERY_USED",
-    "MCP_KEY_RESULTS",
-    "MCP_KEY_SUMMARY",
-    "MCP_LOG_FORMAT",
-    "MCP_LOG_LEVEL_INFO",
-    "MCP_NOT_AVAILABLE",
-    "MCP_PAGINATION_HEADER",
-    "MCP_PROJECT_DELETED",
-    "MCP_SEMANTIC_NOT_AVAILABLE_RESPONSE",
-    "MCP_SERVER_NAME",
-    "MCP_TOOL_EXEC_ERROR",
-    "MCP_TOOL_NAME_QUERY",
-    "MCP_UNKNOWN_TOOL_ERROR",
-    "MCP_UPDATE_ERROR",
-    "MCP_UPDATE_SUCCESS",
-    "MCP_WIPE_CANCELLED",
-    "MCP_WIPE_ERROR",
-    "MCP_WIPE_SUCCESS",
-    "MCP_WRITE_SUCCESS",
-    "METHOD_FIND_WITH_PREFIX",
-    "METHOD_ITEMS",
-    "MIME_TYPE_DEFAULT",
-    "MIME_TYPE_FALLBACK",
-    "MIME_TYPE_PDF",
-    "ML_DEPENDENCIES",
-    "MODEL_COMMAND_PREFIX",
-    "MODEL_CONTEXT_WINDOWS",
-    "MODULE_INDEX_FILE_STEMS",
-    "MODULE_QDRANT_CLIENT",
-    "MODULE_TORCH",
-    "MODULE_TRANSFORMERS",
-    "MOD_RS",
-    "MSG_CHAT_INSTRUCTIONS",
-    "MSG_CONNECTED_MEMGRAPH",
-    "MSG_DOC_NO_CANDIDATES",
-    "MSG_DOC_NO_CONTENT",
-    "MSG_SEMANTIC_NO_RESULTS",
-    "MSG_SEMANTIC_RESULT_FOOTER",
-    "MSG_SEMANTIC_RESULT_HEADER",
-    "MSG_SEMANTIC_SOURCE_FORMAT",
-    "MSG_SEMANTIC_SOURCE_UNAVAILABLE",
-    "MSG_SURGICAL_FAILED",
-    "MSG_SURGICAL_SUCCESS",
-    "MSG_SYNCING_KNOWLEDGE_GRAPH",
-    "MSG_SYNCING_WORKSPACE",
-    "MSG_THINKING_CANCELLED",
-    "MSG_TIMEOUT_FORMAT",
-    "MSG_TOOL_CALL_CANCELLED",
-    "MULTILINE_INPUT_HINT",
-    "MULTIMODAL_EXTENSIONS",
-    "ModelRole",
-    "NAME_FIELDS",
-    "NODE_PROJECT",
-    "NODE_UNIQUE_CONSTRAINTS",
-    "NO_ANSWERS",
-    "NamedTuple",
-    "NodeLabel",
-    "OLLAMA_HEALTH_PATH",
-    "ONEOF_CLASS",
-    "ONEOF_ENUM",
-    "ONEOF_EXTERNAL_MODULE",
-    "ONEOF_EXTERNAL_PACKAGE",
-    "ONEOF_FILE",
-    "ONEOF_FOLDER",
-    "ONEOF_FUNCTION",
-    "ONEOF_INTERFACE",
-    "ONEOF_METHOD",
-    "ONEOF_MODULE",
-    "ONEOF_MODULE_IMPLEMENTATION",
-    "ONEOF_MODULE_INTERFACE",
-    "ONEOF_PACKAGE",
-    "ONEOF_PROJECT",
-    "ONEOF_TYPE",
-    "ONEOF_UNION",
-    "OPENAI_DEFAULT_ENDPOINT",
-    "OPERATOR_PREFIX",
-    "OPTIMIZATION_TABLE_TITLE",
-    "PACKAGE_NAME",
-    "PATH_CURRENT_DIR",
-    "PATH_PARENT_DIR",
-    "PATH_PARENT_PREFIX",
-    "PATH_RELATIVE_PREFIX",
-    "PAYLOAD_NODE_ID",
-    "PAYLOAD_QUALIFIED_NAME",
-    "PERMISSION_MODE_NORMAL_LABEL",
-    "PERMISSION_MODE_TOGGLED",
-    "PERMISSION_MODE_YOLO_LABEL",
-    "PHP_EXTENSIONS",
-    "PKG_CARGO_TOML",
-    "PKG_CMAKE_LISTS",
-    "PKG_CONANFILE",
-    "PKG_INIT_PY",
-    "PKG_MAKEFILE",
-    "PKG_VCXPROJ_GLOB",
-    "PREFIX_ANONYMOUS",
-    "PREFIX_ARROW",
-    "PREFIX_FUNC",
-    "PREFIX_IIFE",
-    "PREFIX_IIFE_DIRECT",
-    "PREFIX_LAMBDA",
-    "PROMPT_ASK_QUESTION",
-    "PROMPT_YOUR_RESPONSE",
-    "PROPERTY_DECORATORS",
-    "PROTOBUF_INDEX_FILE",
-    "PROTOBUF_NODES_FILE",
-    "PROTOBUF_RELS_FILE",
-    "PROTOCOL_BASE_QNS",
-    "PUNCTUATION_TYPES",
-    "PYINSTALLER_ARG_CLEAN",
-    "PYINSTALLER_ARG_COLLECT_ALL",
-    "PYINSTALLER_ARG_COLLECT_DATA",
-    "PYINSTALLER_ARG_COPY_METADATA",
-    "PYINSTALLER_ARG_EXCLUDE_MODULE",
-    "PYINSTALLER_ARG_HIDDEN_IMPORT",
-    "PYINSTALLER_ARG_NAME",
-    "PYINSTALLER_ARG_NOCONFIRM",
-    "PYINSTALLER_ARG_ONEFILE",
-    "PYINSTALLER_CMD",
-    "PYINSTALLER_ENTRY_POINT",
-    "PYINSTALLER_EXCLUDED_MODULES",
-    "PYINSTALLER_PACKAGES",
-    "PYPROJECT_KEY_PACKAGE_DIR",
-    "PYPROJECT_KEY_SETUPTOOLS",
-    "PYPROJECT_KEY_TOOL",
-    "PYPROJECT_PATH",
-    "PY_ASSIGNMENT_QUERY",
-    "PY_BUILTIN_GETATTR",
-    "PY_BUILTIN_LEN",
-    "PY_CLASS_REPOSITORY",
-    "PY_CLS_PREFIX",
-    "PY_DUNDER_BOOL",
-    "PY_DUNDER_CONTAINS",
-    "PY_DUNDER_GETITEM",
-    "PY_DUNDER_LEN",
-    "PY_DUNDER_SETITEM",
-    "PY_EXTENSIONS",
-    "PY_KEYWORD_CLS",
-    "PY_KEYWORD_SELF",
-    "PY_METHOD_CREATE",
-    "PY_METHOD_INIT",
-    "PY_MODELS_BASE_PATH",
-    "PY_NAME_DUNDER",
-    "PY_NAME_UNDERSCORE",
-    "PY_NONE",
-    "PY_OPERAND_REJECT_CHARS",
-    "PY_OP_IN",
-    "PY_PROTOCOL",
-    "PY_RETURN_QUERY",
-    "PY_SCORE_CONTAINS_BASE",
-    "PY_SCORE_EXACT_MATCH",
-    "PY_SCORE_SUFFIX_MATCH",
-    "PY_SELF_PREFIX",
-    "PY_UNION_SEPARATOR",
-    "PY_VAR_PATTERN_ALL",
-    "PY_VAR_SUFFIX_PLURAL",
-    "PermissionMode",
-    "Provider",
-    "PyInstallerPackage",
-    "QUERY_CALLS",
-    "QUERY_CAPTURE_CLASS",
-    "QUERY_CAPTURE_FUNCTION",
-    "QUERY_CLASSES",
-    "QUERY_CONFIG",
-    "QUERY_FUNCTIONS",
-    "QUERY_IMPORTS",
-    "QUERY_KEY_CLASSES",
-    "QUERY_KEY_FUNCTIONS",
-    "QUERY_LANGUAGE",
-    "QUERY_LOCALS",
-    "QUERY_NOT_AVAILABLE",
-    "QUERY_RESULTS_PANEL_TITLE",
-    "QUERY_SUMMARY_DB_ERROR",
-    "QUERY_SUMMARY_SUCCESS",
-    "QUERY_SUMMARY_TIMEOUT",
-    "QUERY_SUMMARY_TRANSLATION_FAILED",
-    "QUERY_SUMMARY_TRUNCATED",
-    "QUOTE_CHARS",
-    "QueryFormat",
-    "REALTIME_LOGGER_FORMAT",
-    "REGEX_FINAL_METHOD_CAPTURE",
-    "REGEX_METHOD_CHAIN_SUFFIX",
-    "REL_TYPE_CALLS",
-    "ROOT_DIR_KEY",
-    "RS_CALL_OR_GENERIC_FN",
-    "RS_COMMENT_TYPES",
-    "RS_DEREF_WRAPPERS",
-    "RS_ENCODING_UTF8",
-    "RS_EXTENSIONS",
-    "RS_FIELD_ARGUMENT",
-    "RS_GUARD_ACCESSORS",
-    "RS_GUARD_WRAPPERS",
-    "RS_IDENTIFIER_TYPES",
-    "RS_IDENTITY_METHODS",
-    "RS_IDENT_NODE_TYPES",
-    "RS_IDENT_OR_SELF",
-    "RS_MACRO_EXPORT_ATTR",
-    "RS_MACRO_RECEIVER_TYPES",
-    "RS_PATH_KEYWORDS",
-    "RS_RESULT_WRAPPERS",
-    "RS_RETURN_STRIP_WRAPPERS",
-    "RS_RETURN_TYPE_NODE_TYPES",
-    "RS_SCOPED_TYPES",
-    "RS_SELF_TYPE",
-    "RS_TYPE_NODE_TYPES",
-    "RS_USE_LIST_DELIMITERS",
-    "RS_WILDCARD_PREFIX",
-    "RUST_CRATE_PREFIX",
-    "RUST_ROOT_FUNCTION_NAMES",
-    "RUST_TRAIT_METHOD_NAMES",
-    "RelationshipType",
-    "SCALA_EXTENSIONS",
-    "SCHEMA_OPTIONAL_SUFFIX",
-    "SCHEMA_PROPS_BRACES",
-    "SEMANTIC_BATCH_SIZE",
-    "SEMANTIC_DEPENDENCIES",
-    "SEMANTIC_TYPE_UNKNOWN",
-    "SEPARATOR_COLON",
-    "SEPARATOR_COMMA",
-    "SEPARATOR_COMMA_SPACE",
-    "SEPARATOR_DOT",
-    "SEPARATOR_DOUBLE_COLON",
-    "SEPARATOR_PROTOTYPE",
-    "SEPARATOR_SLASH",
-    "SESSION_CONTEXT_END",
-    "SESSION_CONTEXT_START",
-    "SESSION_LOG_EXT",
-    "SESSION_LOG_HEADER",
-    "SESSION_LOG_PREFIX",
-    "SESSION_PREFIX_ASSISTANT",
-    "SESSION_PREFIX_USER",
-    "SETUP_PY",
-    "SHELL_CMD_GIT",
-    "SHELL_CMD_GREP",
-    "SHELL_CMD_RM",
-    "SHELL_CMD_WHERE",
-    "SHELL_CMD_WHICH",
-    "SHELL_DANGEROUS_COMMANDS",
-    "SHELL_DANGEROUS_PATTERNS_PIPELINE",
-    "SHELL_DANGEROUS_PATTERNS_SEGMENT",
-    "SHELL_PIPE_OPERATORS",
-    "SHELL_REDIRECT_OPERATORS",
-    "SHELL_RETURN_CODE_ERROR",
-    "SHELL_RM_DANGEROUS_FLAGS",
-    "SHELL_RM_FORCE_FLAG",
-    "SHELL_RM_RF_FLAG",
-    "SHELL_SUBSHELL_PATTERNS",
-    "SHELL_SYSTEM_DIRECTORIES",
-    "SHIFT_TAB_ESCAPE",
-    "SPEC_CPP_CALL_TYPES",
-    "SPEC_CPP_CLASS_TYPES",
-    "SPEC_CPP_FUNCTION_TYPES",
-    "SPEC_CPP_MODULE_TYPES",
-    "SPEC_CPP_PACKAGE_INDICATORS",
-    "SPEC_C_CALL_TYPES",
-    "SPEC_C_CLASS_TYPES",
-    "SPEC_C_FUNCTION_TYPES",
-    "SPEC_C_MODULE_TYPES",
-    "SPEC_C_PACKAGE_INDICATORS",
-    "SPEC_GO_CALL_TYPES",
-    "SPEC_GO_CLASS_TYPES",
-    "SPEC_GO_FUNCTION_TYPES",
-    "SPEC_GO_IMPORT_TYPES",
-    "SPEC_GO_MODULE_TYPES",
-    "SPEC_JAVA_CALL_TYPES",
-    "SPEC_JAVA_CLASS_TYPES",
-    "SPEC_JAVA_FUNCTION_TYPES",
-    "SPEC_JAVA_IMPORT_TYPES",
-    "SPEC_JAVA_MODULE_TYPES",
-    "SPEC_JS_CALL_TYPES",
-    "SPEC_JS_MODULE_TYPES",
-    "SPEC_LUA_CALL_TYPES",
-    "SPEC_LUA_CLASS_TYPES",
-    "SPEC_LUA_FUNCTION_TYPES",
-    "SPEC_LUA_IMPORT_TYPES",
-    "SPEC_LUA_MODULE_TYPES",
-    "SPEC_PHP_CALL_TYPES",
-    "SPEC_PHP_CLASS_TYPES",
-    "SPEC_PHP_FUNCTION_TYPES",
-    "SPEC_PHP_IMPORT_FROM_TYPES",
-    "SPEC_PHP_IMPORT_TYPES",
-    "SPEC_PHP_MODULE_TYPES",
-    "SPEC_PY_CALL_TYPES",
-    "SPEC_PY_CLASS_TYPES",
-    "SPEC_PY_FUNCTION_TYPES",
-    "SPEC_PY_IMPORT_FROM_TYPES",
-    "SPEC_PY_IMPORT_TYPES",
-    "SPEC_PY_MODULE_TYPES",
-    "SPEC_PY_PACKAGE_INDICATORS",
-    "SPEC_RS_CALL_TYPES",
-    "SPEC_RS_CLASS_TYPES",
-    "SPEC_RS_FUNCTION_TYPES",
-    "SPEC_RS_IMPORT_FROM_TYPES",
-    "SPEC_RS_IMPORT_TYPES",
-    "SPEC_RS_MODULE_TYPES",
-    "SPEC_RS_PACKAGE_INDICATORS",
-    "SPEC_SCALA_CALL_TYPES",
-    "SPEC_SCALA_CLASS_TYPES",
-    "SPEC_SCALA_FUNCTION_TYPES",
-    "SPEC_SCALA_IMPORT_TYPES",
-    "SPEC_SCALA_MODULE_TYPES",
-    "STATUS_BAR_BRANCH_CLEAN_HTML",
-    "STATUS_BAR_BRANCH_CLEAN_PLAIN",
-    "STATUS_BAR_BRANCH_DIRTY_HTML",
-    "STATUS_BAR_BRANCH_DIRTY_PLAIN",
-    "STATUS_BAR_BRANCH_RICH_TEXT",
-    "STATUS_BAR_CLEAN_STYLE",
-    "STATUS_BAR_CONFIG_COLOR",
-    "STATUS_BAR_CONFIG_LABEL_C",
-    "STATUS_BAR_CONFIG_LABEL_COLOR",
-    "STATUS_BAR_CONFIG_LABEL_EDIT",
-    "STATUS_BAR_CONFIG_LABEL_INSTRUCTIONS",
-    "STATUS_BAR_CONFIG_LABEL_O",
-    "STATUS_BAR_CONFIG_LABEL_REPO",
-    "STATUS_BAR_CONFIG_SEPARATOR",
-    "STATUS_BAR_DIRTY_MARKER",
-    "STATUS_BAR_DIRTY_STYLE",
-    "STATUS_BAR_EDIT_OFF",
-    "STATUS_BAR_EDIT_ON",
-    "STATUS_BAR_SEPARATOR_CHAR",
-    "STATUS_BAR_SEPARATOR_COLOR",
-    "STATUS_BAR_SPINNER",
-    "STATUS_BAR_TOKEN_HTML",
-    "STR_NONE",
-    "StrEnum",
-    "StyleModifier",
-    "SupportedLanguage",
-    "TABLE_COL_CONFIGURATION",
-    "TABLE_COL_VALUE",
-    "TABLE_ROW_CYPHER_MODEL",
-    "TABLE_ROW_EDIT_CONFIRMATION",
-    "TABLE_ROW_OLLAMA_CYPHER",
-    "TABLE_ROW_OLLAMA_ENDPOINT",
-    "TABLE_ROW_OLLAMA_ORCHESTRATOR",
-    "TABLE_ROW_ORCHESTRATOR_MODEL",
-    "TABLE_ROW_TARGET_LANGUAGE",
-    "TABLE_ROW_TARGET_REPOSITORY",
-    "TEST_PATH_PATTERNS",
-    "TEXT_UNKNOWN",
-    "TIKTOKEN_ENCODING",
-    "TMP_DIR",
-    "TMP_EXTENSION",
-    "TOKEN_COLOR_CRITICAL",
-    "TOKEN_COLOR_OK",
-    "TOKEN_COLOR_WARNING",
-    "TOKEN_THRESHOLD_CRITICAL",
-    "TOKEN_THRESHOLD_WARNING",
-    "TOML_KEY_OPTIONAL_DEPS",
-    "TOML_KEY_PROJECT",
-    "TREESITTER_EXTRA_KEY",
-    "TREESITTER_PKG_PREFIX",
-    "TREE_SITTER_MODULE_PREFIX",
-    "TREE_SITTER_PREFIX",
-    "TRIE_INTERNAL_PREFIX",
-    "TRIE_QN_KEY",
-    "TRIE_TYPE_KEY",
-    "TRIPLE_QUOTES",
-    "TSCONFIG_FILENAMES",
-    "TSX_EXTENSIONS",
-    "TS_ABSTRACT_CLASS_DECLARATION",
-    "TS_ACCESSIBILITY_MODIFIER",
-    "TS_ACCESS_SPECIFIER",
-    "TS_ALIASED_IMPORT",
-    "TS_ALIAS_SKIP_DIRS",
-    "TS_ANNOTATION",
-    "TS_ANNOTATION_TYPE_DECLARATION",
-    "TS_ARGUMENTS",
-    "TS_ARRAY",
-    "TS_ARRAY_CREATION_EXPRESSION",
-    "TS_ARROW_FUNCTION",
-    "TS_ASSIGNMENT_EXPRESSION",
-    "TS_ASTERISK",
-    "TS_AS_EXPRESSION",
-    "TS_ATTRIBUTE",
-    "TS_BASE_CLASS_CLAUSE",
-    "TS_BASE_URL_KEY",
-    "TS_BINARY_EXPRESSION",
-    "TS_CALL_EXPRESSION",
-    "TS_CAST_WRAPPER_TYPES",
-    "TS_CLASS_BODY",
-    "TS_CLASS_DECLARATION",
-    "TS_CLASS_HERITAGE",
-    "TS_COMPILER_OPTIONS_KEY",
-    "TS_CONSTRUCTOR_DECLARATION",
-    "TS_CPP_BINARY_EXPRESSION",
-    "TS_CPP_CALL_EXPRESSION",
-    "TS_CPP_DECLARATION",
-    "TS_CPP_DELETE_EXPRESSION",
-    "TS_CPP_FIELD_DECLARATION",
-    "TS_CPP_FIELD_EXPRESSION",
-    "TS_CPP_FUNCTION_DECLARATOR",
-    "TS_CPP_FUNCTION_DEFINITION",
-    "TS_CPP_LAMBDA_EXPRESSION",
-    "TS_CPP_LINKAGE_SPECIFICATION",
-    "TS_CPP_NEW_EXPRESSION",
-    "TS_CPP_SUBSCRIPT_EXPRESSION",
-    "TS_CPP_TEMPLATE_DECLARATION",
-    "TS_CPP_TEMPLATE_PARAMETER_LIST",
-    "TS_CPP_TRANSLATION_UNIT",
-    "TS_CPP_UNARY_EXPRESSION",
-    "TS_CPP_UPDATE_EXPRESSION",
-    "TS_DECIMAL_FLOATING_POINT_LITERAL",
-    "TS_DECLARATION",
-    "TS_DECORATOR",
-    "TS_DOTTED_NAME",
-    "TS_DOT_INDEX_EXPRESSION",
-    "TS_ENHANCED_FOR_STATEMENT",
-    "TS_ENUM_CLASS_SPECIFIER",
-    "TS_ENUM_DECLARATION",
-    "TS_ENUM_SPECIFIER",
-    "TS_ERROR",
-    "TS_EXPORT_CLAUSE",
-    "TS_EXPORT_DEFAULT",
-    "TS_EXPORT_SPECIFIER",
-    "TS_EXPORT_STATEMENT",
-    "TS_EXPRESSION_STATEMENT",
-    "TS_EXTENDS",
-    "TS_EXTENDS_CLAUSE",
-    "TS_EXTENDS_TYPE_CLAUSE",
-    "TS_EXTENSIONS",
-    "TS_FALSE",
-    "TS_FIELD_ACCESS",
-    "TS_FIELD_ARGUMENT",
-    "TS_FIELD_ARGUMENTS",
-    "TS_FIELD_BODY",
-    "TS_FIELD_CONDITION",
-    "TS_FIELD_DECLARATION",
-    "TS_FIELD_DECLARATOR",
-    "TS_FIELD_FUNCTION",
-    "TS_FIELD_INTERFACES",
-    "TS_FIELD_LEFT",
-    "TS_FIELD_NAME",
-    "TS_FIELD_OBJECT",
-    "TS_FIELD_OPERATORS",
-    "TS_FIELD_PARAMETER",
-    "TS_FIELD_PARAMETERS",
-    "TS_FIELD_PATTERN",
-    "TS_FIELD_RIGHT",
-    "TS_FIELD_SUPERCLASS",
-    "TS_FIELD_TYPE",
-    "TS_FIELD_TYPE_PARAMETERS",
-    "TS_FORMAL_PARAMETER",
-    "TS_FUNCTION_DECLARATION",
-    "TS_FUNCTION_EXPRESSION",
-    "TS_FUNCTION_SIGNATURE",
-    "TS_GENERATOR_FUNCTION_DECLARATION",
-    "TS_GENERIC_FUNCTION",
-    "TS_GENERIC_TYPE",
-    "TS_GO_ASSIGNMENT_STATEMENT",
-    "TS_GO_CALL_EXPRESSION",
-    "TS_GO_COMPOSITE_LITERAL",
-    "TS_GO_CONTAINER_TYPES",
-    "TS_GO_EXPRESSION_LIST",
-    "TS_GO_FIELD_DECLARATION",
-    "TS_GO_FIELD_DECLARATION_LIST",
-    "TS_GO_FIELD_IDENTIFIER",
-    "TS_GO_FUNCTION_DECLARATION",
-    "TS_GO_FUNC_LITERAL",
-    "TS_GO_IMPORT_DECLARATION",
-    "TS_GO_INTERFACE_TYPE",
-    "TS_GO_KEYED_ELEMENT",
-    "TS_GO_LITERAL_ELEMENT",
-    "TS_GO_LITERAL_VALUE",
-    "TS_GO_METHOD_DECLARATION",
-    "TS_GO_PARAMETER_DECLARATION",
-    "TS_GO_PARAMETER_LIST",
-    "TS_GO_POINTER_TYPE",
-    "TS_GO_SELECTOR_EXPRESSION",
-    "TS_GO_SHORT_VAR_DECLARATION",
-    "TS_GO_SOURCE_FILE",
-    "TS_GO_STRUCT_TYPE",
-    "TS_GO_TYPE_ALIAS",
-    "TS_GO_TYPE_ASSERTION_EXPRESSION",
-    "TS_GO_TYPE_DECLARATION",
-    "TS_GO_TYPE_SPEC",
-    "TS_GO_UNARY_EXPRESSION",
-    "TS_GO_VAR_DECLARATION",
-    "TS_GO_VAR_SPEC",
-    "TS_IDENTIFIER",
-    "TS_IMPLEMENTS_CLAUSE",
-    "TS_IMPL_ITEM",
-    "TS_IMPORT_CLAUSE",
-    "TS_IMPORT_DECLARATION",
-    "TS_IMPORT_FROM_STATEMENT",
-    "TS_IMPORT_PREFIX",
-    "TS_IMPORT_SPEC",
-    "TS_IMPORT_SPECIFIER",
-    "TS_IMPORT_SPEC_LIST",
-    "TS_IMPORT_STATEMENT",
-    "TS_INTEGER_LITERAL",
-    "TS_INTERFACE_DECLARATION",
-    "TS_INTERNAL_MODULE",
-    "TS_INTERPRETED_STRING_LITERAL",
-    "TS_JAVA_ANNOTATION_TYPE_DECLARATION",
-    "TS_JAVA_CAST_EXPRESSION",
-    "TS_JAVA_EXTENDS_INTERFACES",
-    "TS_JAVA_METHOD_INVOCATION",
-    "TS_JSX_EXPRESSION",
-    "TS_JSX_OPENING_ELEMENT",
-    "TS_JSX_SELF_CLOSING_ELEMENT",
-    "TS_LEXICAL_DECLARATION",
-    "TS_LOCALS_PATTERN",
-    "TS_LOCAL_VARIABLE_DECLARATION",
-    "TS_LUA_ASSIGNMENT_STATEMENT",
-    "TS_LUA_CHUNK",
-    "TS_LUA_EXPRESSION_LIST",
-    "TS_LUA_FUNCTION_CALL",
-    "TS_LUA_FUNCTION_DECLARATION",
-    "TS_LUA_FUNCTION_DEFINITION",
-    "TS_LUA_IDENTIFIER",
-    "TS_LUA_LOCAL_STATEMENT",
-    "TS_LUA_METHOD_INDEX_EXPRESSION",
-    "TS_LUA_VARIABLE_DECLARATION",
-    "TS_LUA_VARIABLE_LIST",
-    "TS_MARKER_ANNOTATION",
-    "TS_MEMBER_EXPRESSION",
-    "TS_METHOD_DECLARATION",
-    "TS_METHOD_DEFINITION",
-    "TS_METHOD_INVOCATION",
-    "TS_MODIFIERS",
-    "TS_MODULE",
-    "TS_MODULE_DECLARATION",
-    "TS_NAMED_IMPORTS",
-    "TS_NAMESPACE_DEFINITION",
-    "TS_NAMESPACE_IMPORT",
-    "TS_NEW_EXPRESSION",
-    "TS_NON_NULL_EXPRESSION",
-    "TS_OBJECT",
-    "TS_OBJECT_CREATION_EXPRESSION",
-    "TS_OBJECT_PATTERN",
-    "TS_OPTIONAL_PARAMETER",
-    "TS_PACKAGE_DECLARATION",
-    "TS_PACKAGE_IDENTIFIER",
-    "TS_PAIR",
-    "TS_PAIR_PATTERN",
-    "TS_PARENTHESIZED_EXPRESSION",
-    "TS_PATHS_KEY",
-    "TS_PHP_ANONYMOUS_FUNCTION",
-    "TS_PHP_ARROW_FUNCTION",
-    "TS_PHP_ATTRIBUTE",
-    "TS_PHP_ATTRIBUTE_GROUP",
-    "TS_PHP_ATTRIBUTE_LIST",
-    "TS_PHP_BASE_CLAUSE",
-    "TS_PHP_CLASS_INTERFACE_CLAUSE",
-    "TS_PHP_FUNCTION",
-    "TS_PHP_FUNCTION_CALL_EXPRESSION",
-    "TS_PHP_FUNCTION_DEFINITION",
-    "TS_PHP_FUNCTION_STATIC_DECLARATION",
-    "TS_PHP_INCLUDE_EXPRESSION",
-    "TS_PHP_INCLUDE_ONCE_EXPRESSION",
-    "TS_PHP_MEMBER_CALL_EXPRESSION",
-    "TS_PHP_METHOD_DECLARATION",
-    "TS_PHP_NAME",
-    "TS_PHP_NAMESPACE_DEFINITION",
-    "TS_PHP_NAMESPACE_USE_CLAUSE",
-    "TS_PHP_NAMESPACE_USE_DECLARATION",
-    "TS_PHP_NULLSAFE_MEMBER_CALL_EXPRESSION",
-    "TS_PHP_OBJECT_CREATION_EXPRESSION",
-    "TS_PHP_QUALIFIED_NAME",
-    "TS_PHP_REQUIRE_EXPRESSION",
-    "TS_PHP_REQUIRE_ONCE_EXPRESSION",
-    "TS_PHP_SCOPED_CALL_EXPRESSION",
-    "TS_PHP_TRAIT_DECLARATION",
-    "TS_PHP_USE_DECLARATION",
-    "TS_PHP_VISIBILITY_MODIFIER",
-    "TS_PREPROC_INCLUDE",
-    "TS_PRIVATE",
-    "TS_PRIVATE_PROPERTY_IDENTIFIER",
-    "TS_PROGRAM",
-    "TS_PROPERTY_IDENTIFIER",
-    "TS_PY_ASSIGNMENT",
-    "TS_PY_ATTRIBUTE",
-    "TS_PY_BLOCK",
-    "TS_PY_BOOLEAN_OPERATOR",
-    "TS_PY_CALL",
-    "TS_PY_CLASS_DEFINITION",
-    "TS_PY_COMPARISON_OPERATOR",
-    "TS_PY_CONDITIONAL_EXPRESSION",
-    "TS_PY_DECORATED_DEFINITION",
-    "TS_PY_DECORATOR",
-    "TS_PY_DEFAULT_PARAMETER",
-    "TS_PY_DICTIONARY",
-    "TS_PY_DICTIONARY_SPLAT_PATTERN",
-    "TS_PY_ELIF_CLAUSE",
-    "TS_PY_EXPRESSION_STATEMENT",
-    "TS_PY_FOR_IN_CLAUSE",
-    "TS_PY_FOR_STATEMENT",
-    "TS_PY_FUNCTION_DEFINITION",
-    "TS_PY_IDENTIFIER",
-    "TS_PY_IF_STATEMENT",
-    "TS_PY_IMPORT_FROM_STATEMENT",
-    "TS_PY_IMPORT_STATEMENT",
-    "TS_PY_KEYWORD",
-    "TS_PY_KEYWORD_ARGUMENT",
-    "TS_PY_LAMBDA",
-    "TS_PY_LIST",
-    "TS_PY_LIST_COMPREHENSION",
-    "TS_PY_LIST_SPLAT_PATTERN",
-    "TS_PY_MODULE",
-    "TS_PY_NOT_OPERATOR",
-    "TS_PY_PAIR",
-    "TS_PY_RETURN",
-    "TS_PY_RETURN_STATEMENT",
-    "TS_PY_SET",
-    "TS_PY_STRING",
-    "TS_PY_STRING_CONTENT",
-    "TS_PY_SUBSCRIPT",
-    "TS_PY_TRY_STATEMENT",
-    "TS_PY_TUPLE",
-    "TS_PY_TYPED_DEFAULT_PARAMETER",
-    "TS_PY_TYPED_PARAMETER",
-    "TS_PY_WHILE_STATEMENT",
-    "TS_PY_WITH_STATEMENT",
-    "TS_RECORD_DECLARATION",
-    "TS_RELATIVE_IMPORT",
-    "TS_REQUIRED_PARAMETER",
-    "TS_RETURN",
-    "TS_RETURN_STATEMENT",
-    "TS_RS_ATTRIBUTE_ITEM",
-    "TS_RS_BLOCK_COMMENT",
-    "TS_RS_CALL_EXPRESSION",
-    "TS_RS_CLOSURE_EXPRESSION",
-    "TS_RS_CRATE",
-    "TS_RS_ENUM_ITEM",
-    "TS_RS_EXTERN_CRATE_DECLARATION",
-    "TS_RS_FIELD_DECLARATION",
-    "TS_RS_FIELD_DECLARATION_LIST",
-    "TS_RS_FIELD_EXPRESSION",
-    "TS_RS_FIELD_IDENTIFIER",
-    "TS_RS_FIELD_PATH",
-    "TS_RS_FUNCTION_ITEM",
-    "TS_RS_FUNCTION_SIGNATURE_ITEM",
-    "TS_RS_IMPL_ITEM",
-    "TS_RS_INNER_ATTRIBUTE_ITEM",
-    "TS_RS_KEYWORD_AS",
-    "TS_RS_LET_DECLARATION",
-    "TS_RS_LINE_COMMENT",
-    "TS_RS_MACRO_DEFINITION",
-    "TS_RS_MACRO_INVOCATION",
-    "TS_RS_MATCH_ARM",
-    "TS_RS_MATCH_EXPRESSION",
-    "TS_RS_MOD_ITEM",
-    "TS_RS_PARAMETER",
-    "TS_RS_POINTER_TYPE",
-    "TS_RS_PRIMITIVE_TYPE",
-    "TS_RS_REFERENCE_TYPE",
-    "TS_RS_SCOPED_TYPE_IDENTIFIER",
-    "TS_RS_SCOPED_USE_LIST",
-    "TS_RS_SELF_PARAMETER",
-    "TS_RS_SOURCE_FILE",
-    "TS_RS_STRUCT_EXPRESSION",
-    "TS_RS_STRUCT_ITEM",
-    "TS_RS_TOKEN_DOT",
-    "TS_RS_TRAIT_BOUNDS",
-    "TS_RS_TRAIT_ITEM",
-    "TS_RS_TRY_EXPRESSION",
-    "TS_RS_TUPLE_STRUCT_PATTERN",
-    "TS_RS_TYPE_ARGUMENTS",
-    "TS_RS_TYPE_ITEM",
-    "TS_RS_UNION_ITEM",
-    "TS_RS_USE_AS_CLAUSE",
-    "TS_RS_USE_DECLARATION",
-    "TS_RS_USE_LIST",
-    "TS_RS_USE_WILDCARD",
-    "TS_SATISFIES_EXPRESSION",
-    "TS_SCALA_CALL_EXPRESSION",
-    "TS_SCALA_CLASS_DEFINITION",
-    "TS_SCALA_COMPILATION_UNIT",
-    "TS_SCALA_FIELD_EXPRESSION",
-    "TS_SCALA_FUNCTION_DECLARATION",
-    "TS_SCALA_FUNCTION_DEFINITION",
-    "TS_SCALA_GENERIC_FUNCTION",
-    "TS_SCALA_IMPORT_DECLARATION",
-    "TS_SCALA_INFIX_EXPRESSION",
-    "TS_SCALA_OBJECT_DEFINITION",
-    "TS_SCALA_TRAIT_DEFINITION",
-    "TS_SCOPED_IDENTIFIER",
-    "TS_SCOPED_TYPE_IDENTIFIER",
-    "TS_SELECTOR_EXPRESSION",
-    "TS_SHORTHAND_PROPERTY_IDENTIFIER_PATTERN",
-    "TS_SPREAD_PARAMETER",
-    "TS_STATEMENT_BLOCK",
-    "TS_STATIC",
-    "TS_STRING",
-    "TS_STRING_LITERAL",
-    "TS_STRUCT_SPECIFIER",
-    "TS_SUPER",
-    "TS_SYSTEM_LIB_STRING",
-    "TS_TEMPLATE_ARGUMENT_LIST",
-    "TS_TEMPLATE_FUNCTION",
-    "TS_TEMPLATE_TYPE",
-    "TS_THIS",
-    "TS_TRUE",
-    "TS_TYPE_ALIAS_DECLARATION",
-    "TS_TYPE_DESCRIPTOR",
-    "TS_TYPE_IDENTIFIER",
-    "TS_TYPE_LIST",
-    "TS_TYPE_PARAMETER",
-    "TS_UNION_SPECIFIER",
-    "TS_USE_DECLARATION",
-    "TS_VARIABLE_DECLARATION",
-    "TS_VARIABLE_DECLARATOR",
-    "TS_VIRTUAL",
-    "TS_VOID_TYPE",
-    "TS_WILDCARD_IMPORT",
-    "TYPE_INFERENCE_BASE_MODEL",
-    "TYPE_INFERENCE_LIST",
-    "TreeSitterModule",
-    "UI_DIFF_FILE_HEADER",
-    "UI_ERR_EXPORT_FAILED",
-    "UI_ERR_UNEXPECTED",
-    "UI_FEEDBACK_PROMPT",
-    "UI_GRAPH_EXPORT_STATS",
-    "UI_GRAPH_EXPORT_SUCCESS",
-    "UI_HELP_COMMANDS",
-    "UI_INPUT_PROMPT_HTML",
-    "UI_MODEL_CURRENT",
-    "UI_MODEL_SWITCHED",
-    "UI_MODEL_SWITCH_ERROR",
-    "UI_MODEL_USAGE",
-    "UI_NEW_FILE_HEADER",
-    "UI_OPTIMIZATION_INIT",
-    "UI_OPTIMIZATION_PANEL",
-    "UI_OPTIMIZATION_START",
-    "UI_REFERENCE_DOC_INFO",
-    "UI_SHELL_COMMAND_HEADER",
-    "UI_TOOL_APPROVAL",
-    "UI_TOOL_ARGS_FORMAT",
-    "UNIXCODER_BUFFER_BIAS",
-    "UNIXCODER_MASK_TOKEN",
-    "UNIXCODER_MAX_CONTEXT",
-    "UNIXCODER_MODEL",
-    "UniXcoderMode",
-    "UniqueKeyType",
-    "V1_PATH",
-    "VERSION_SPLIT_EQ",
-    "VERSION_SPLIT_GTE",
-    "VERSION_SPLIT_LT",
-    "WATCHER_SLEEP_INTERVAL",
-    "YES_ANSWER",
-    "YES_ANSWERS",
-]
+from .ast_cpp import *  # noqa: F403
+from .ast_go import *  # noqa: F403
+from .ast_java import *  # noqa: F403
+from .ast_js import *  # noqa: F403
+from .ast_lua import *  # noqa: F403
+from .ast_nodes import *  # noqa: F403
+from .ast_php import *  # noqa: F403
+from .ast_python import *  # noqa: F403
+from .ast_rust import *  # noqa: F403
+from .ast_scala import *  # noqa: F403
+from .build import *  # noqa: F403
+from .cli import *  # noqa: F403
+from .core import *  # noqa: F403
+from .deadcode_roots import *  # noqa: F403
+from .deps import *  # noqa: F403
+from .fqn_specs import *  # noqa: F403
+from .graph import *  # noqa: F403
+from .graph import _NODE_LABEL_UNIQUE_KEYS as _NODE_LABEL_UNIQUE_KEYS
+from .health import *  # noqa: F403
+from .lang_tooling import *  # noqa: F403
+from .languages import *  # noqa: F403
+from .mcp import *  # noqa: F403
+from .providers import *  # noqa: F403
+from .security import *  # noqa: F403
+from .stdlib_types import *  # noqa: F403
