+++ codebase_rag/constants/fqn_specs.py
@@ -1,0 +1,502 @@
+# (H) Per-language FQN scope types and language-spec node tuples.
+
+from .ast_cpp import (
+    TS_CPP_BINARY_EXPRESSION,
+    TS_CPP_CALL_EXPRESSION,
+    TS_CPP_DECLARATION,
+    TS_CPP_DELETE_EXPRESSION,
+    TS_CPP_FIELD_DECLARATION,
+    TS_CPP_FIELD_EXPRESSION,
+    TS_CPP_FUNCTION_DEFINITION,
+    TS_CPP_LAMBDA_EXPRESSION,
+    TS_CPP_LINKAGE_SPECIFICATION,
+    TS_CPP_NEW_EXPRESSION,
+    TS_CPP_SUBSCRIPT_EXPRESSION,
+    TS_CPP_TEMPLATE_DECLARATION,
+    TS_CPP_TRANSLATION_UNIT,
+    TS_CPP_UNARY_EXPRESSION,
+    TS_CPP_UPDATE_EXPRESSION,
+    CppNodeType,
+)
+from .ast_go import (
+    TS_GO_CALL_EXPRESSION,
+    TS_GO_FUNCTION_DECLARATION,
+    TS_GO_IMPORT_DECLARATION,
+    TS_GO_METHOD_DECLARATION,
+    TS_GO_SOURCE_FILE,
+    TS_GO_TYPE_ALIAS,
+    TS_GO_TYPE_DECLARATION,
+    TS_GO_TYPE_SPEC,
+)
+from .ast_java import (
+    TS_CONSTRUCTOR_DECLARATION,
+    TS_JAVA_ANNOTATION_TYPE_DECLARATION,
+    TS_JAVA_METHOD_INVOCATION,
+    TS_METHOD_DECLARATION,
+    TS_OBJECT_CREATION_EXPRESSION,
+    TS_PROGRAM,
+    TS_RECORD_DECLARATION,
+)
+from .ast_js import (
+    TS_ARROW_FUNCTION,
+    TS_CLASS_BODY,
+    TS_FUNCTION_DECLARATION,
+    TS_FUNCTION_EXPRESSION,
+    TS_GENERATOR_FUNCTION_DECLARATION,
+    TS_OBJECT,
+)
+from .ast_lua import (
+    TS_LUA_CHUNK,
+    TS_LUA_FUNCTION_CALL,
+    TS_LUA_FUNCTION_DECLARATION,
+    TS_LUA_FUNCTION_DEFINITION,
+)
+from .ast_nodes import (
+    TS_CALL_EXPRESSION,
+    TS_CLASS_DECLARATION,
+    TS_ENUM_DECLARATION,
+    TS_ENUM_SPECIFIER,
+    TS_FUNCTION_SIGNATURE,
+    TS_IMPORT_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_INTERNAL_MODULE,
+    TS_METHOD_DEFINITION,
+    TS_NAMESPACE_DEFINITION,
+    TS_NEW_EXPRESSION,
+    TS_STATEMENT_BLOCK,
+    TS_STRUCT_SPECIFIER,
+    TS_UNION_SPECIFIER,
+)
+from .ast_php import (
+    TS_PHP_ANONYMOUS_FUNCTION,
+    TS_PHP_ARROW_FUNCTION,
+    TS_PHP_FUNCTION_CALL_EXPRESSION,
+    TS_PHP_FUNCTION_DEFINITION,
+    TS_PHP_INCLUDE_EXPRESSION,
+    TS_PHP_INCLUDE_ONCE_EXPRESSION,
+    TS_PHP_MEMBER_CALL_EXPRESSION,
+    TS_PHP_METHOD_DECLARATION,
+    TS_PHP_NAMESPACE_DEFINITION,
+    TS_PHP_NAMESPACE_USE_DECLARATION,
+    TS_PHP_NULLSAFE_MEMBER_CALL_EXPRESSION,
+    TS_PHP_OBJECT_CREATION_EXPRESSION,
+    TS_PHP_REQUIRE_EXPRESSION,
+    TS_PHP_REQUIRE_ONCE_EXPRESSION,
+    TS_PHP_SCOPED_CALL_EXPRESSION,
+    TS_PHP_TRAIT_DECLARATION,
+)
+from .ast_python import (
+    TS_PY_CALL,
+    TS_PY_CLASS_DEFINITION,
+    TS_PY_FUNCTION_DEFINITION,
+    TS_PY_IMPORT_FROM_STATEMENT,
+    TS_PY_IMPORT_STATEMENT,
+    TS_PY_MODULE,
+    TS_PY_PAIR,
+    TS_PY_WITH_STATEMENT,
+)
+from .ast_rust import (
+    TS_RS_CALL_EXPRESSION,
+    TS_RS_CLOSURE_EXPRESSION,
+    TS_RS_ENUM_ITEM,
+    TS_RS_EXTERN_CRATE_DECLARATION,
+    TS_RS_FUNCTION_ITEM,
+    TS_RS_FUNCTION_SIGNATURE_ITEM,
+    TS_RS_IMPL_ITEM,
+    TS_RS_MACRO_DEFINITION,
+    TS_RS_MACRO_INVOCATION,
+    TS_RS_MOD_ITEM,
+    TS_RS_SOURCE_FILE,
+    TS_RS_STRUCT_ITEM,
+    TS_RS_TRAIT_ITEM,
+    TS_RS_TYPE_ITEM,
+    TS_RS_UNION_ITEM,
+    TS_RS_USE_DECLARATION,
+)
+from .ast_scala import (
+    TS_SCALA_CALL_EXPRESSION,
+    TS_SCALA_CLASS_DEFINITION,
+    TS_SCALA_COMPILATION_UNIT,
+    TS_SCALA_FIELD_EXPRESSION,
+    TS_SCALA_FUNCTION_DECLARATION,
+    TS_SCALA_FUNCTION_DEFINITION,
+    TS_SCALA_GENERIC_FUNCTION,
+    TS_SCALA_IMPORT_DECLARATION,
+    TS_SCALA_INFIX_EXPRESSION,
+    TS_SCALA_OBJECT_DEFINITION,
+    TS_SCALA_TRAIT_DEFINITION,
+)
+from .languages import (
+    PKG_CARGO_TOML,
+    PKG_CMAKE_LISTS,
+    PKG_CONANFILE,
+    PKG_INIT_PY,
+    PKG_MAKEFILE,
+    PKG_VCXPROJ_GLOB,
+)
+
+# (H) FQN node type tuples for Python
+FQN_PY_SCOPE_TYPES = (
+    TS_PY_CLASS_DEFINITION,
+    TS_PY_MODULE,
+    TS_PY_FUNCTION_DEFINITION,
+)
+FQN_PY_FUNCTION_TYPES = (TS_PY_FUNCTION_DEFINITION,)
+
+# (H) FQN node type tuples for JS
+FQN_JS_SCOPE_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_PROGRAM,
+    TS_FUNCTION_DECLARATION,
+    TS_FUNCTION_EXPRESSION,
+    TS_ARROW_FUNCTION,
+    TS_METHOD_DEFINITION,
+)
+FQN_JS_FUNCTION_TYPES = (
+    TS_FUNCTION_DECLARATION,
+    TS_METHOD_DEFINITION,
+    TS_ARROW_FUNCTION,
+    TS_FUNCTION_EXPRESSION,
+)
+
+# (H) FQN node type tuples for TS. The grammar emits `internal_module` for a
+# (H) `namespace`/`module` block; without it a class declared inside a namespace
+# (H) loses the namespace from its qn and collides with a top-level same name.
+FQN_TS_SCOPE_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_NAMESPACE_DEFINITION,
+    TS_INTERNAL_MODULE,
+    TS_PROGRAM,
+    TS_FUNCTION_DECLARATION,
+    TS_FUNCTION_EXPRESSION,
+    TS_ARROW_FUNCTION,
+    TS_METHOD_DEFINITION,
+)
+FQN_TS_FUNCTION_TYPES = (
+    TS_FUNCTION_DECLARATION,
+    TS_METHOD_DEFINITION,
+    TS_ARROW_FUNCTION,
+    TS_FUNCTION_EXPRESSION,
+    TS_FUNCTION_SIGNATURE,
+)
+
+# (H) When climbing a nameless arrow's ancestors to find its binding declarator,
+# (H) crossing one of these means the arrow lives INSIDE another function's body
+# (H) (a JSX event handler, a `.map()` callback), not directly bound to the outer
+# (H) const -- so it must not inherit that const's name. Without this stop the inner
+# (H) arrow becomes `Component.Component`/`utils.getInitials.getInitials`, a
+# (H) double-segment phantom with no incoming edge (dead-code false positive) that
+# (H) also steals the enclosing scope from the real inline handler nodes.
+JS_ARROW_NAME_CLIMB_BOUNDARY = frozenset(
+    {
+        TS_STATEMENT_BLOCK,
+        TS_ARROW_FUNCTION,
+        TS_FUNCTION_DECLARATION,
+        TS_FUNCTION_EXPRESSION,
+        TS_METHOD_DEFINITION,
+        TS_GENERATOR_FUNCTION_DECLARATION,
+        TS_CLASS_BODY,
+        # (H) An OBJECT VALUE (`var pets = { list: function(req, res) {...} }`)
+        # (H) is owned by the pair-key naming path; climbing past it would steal
+        # (H) the enclosing declarator's name and register a phantom function.
+        TS_PY_PAIR,
+        TS_OBJECT,
+    }
+)
+
+# (H) FQN node type tuples for Rust
+FQN_RS_SCOPE_TYPES = (
+    TS_RS_STRUCT_ITEM,
+    TS_RS_ENUM_ITEM,
+    TS_RS_TRAIT_ITEM,
+    TS_RS_IMPL_ITEM,
+    TS_RS_MOD_ITEM,
+    TS_RS_SOURCE_FILE,
+)
+FQN_RS_FUNCTION_TYPES = (
+    TS_RS_FUNCTION_ITEM,
+    TS_RS_FUNCTION_SIGNATURE_ITEM,
+    TS_RS_CLOSURE_EXPRESSION,
+)
+
+# (H) FQN node type tuples for Java
+FQN_JAVA_SCOPE_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_ENUM_DECLARATION,
+    TS_PROGRAM,
+    # (H) An enclosing method/constructor is a qn scope for a function nested in its
+    # (H) body (a method-body anonymous class's method): the call pass builds
+    # (H) `Class.method.nested`, so the definition pass must include the method too --
+    # (H) else the two disagree and every edge from the nested function dangles onto a
+    # (H) phantom `Class.method.nested` node, orphaning its callees. A direct method is
+    # (H) the func itself, not an ancestor, so its own qn is unaffected.
+    TS_METHOD_DECLARATION,
+    TS_CONSTRUCTOR_DECLARATION,
+)
+FQN_JAVA_FUNCTION_TYPES = (
+    TS_METHOD_DECLARATION,
+    TS_CONSTRUCTOR_DECLARATION,
+)
+
+# (H) FQN node type tuples for C++
+FQN_CPP_SCOPE_TYPES = (
+    CppNodeType.CLASS_SPECIFIER,
+    TS_STRUCT_SPECIFIER,
+    TS_NAMESPACE_DEFINITION,
+    TS_CPP_TRANSLATION_UNIT,
+)
+FQN_CPP_FUNCTION_TYPES = (
+    TS_CPP_FUNCTION_DEFINITION,
+    TS_CPP_DECLARATION,
+    TS_CPP_FIELD_DECLARATION,
+    TS_CPP_TEMPLATE_DECLARATION,
+    TS_CPP_LAMBDA_EXPRESSION,
+)
+
+# (H) FQN node type tuples for Lua
+FQN_LUA_SCOPE_TYPES = (TS_LUA_CHUNK,)
+FQN_LUA_FUNCTION_TYPES = (
+    TS_LUA_FUNCTION_DECLARATION,
+    TS_LUA_FUNCTION_DEFINITION,
+)
+
+# (H) FQN node type tuples for Go
+FQN_GO_SCOPE_TYPES = (
+    TS_GO_TYPE_DECLARATION,
+    TS_GO_SOURCE_FILE,
+)
+FQN_GO_FUNCTION_TYPES = (
+    TS_GO_FUNCTION_DECLARATION,
+    TS_GO_METHOD_DECLARATION,
+)
+
+# (H) FQN node type tuples for Scala
+FQN_SCALA_SCOPE_TYPES = (
+    TS_SCALA_CLASS_DEFINITION,
+    TS_SCALA_OBJECT_DEFINITION,
+    TS_SCALA_TRAIT_DEFINITION,
+    TS_SCALA_COMPILATION_UNIT,
+)
+FQN_SCALA_FUNCTION_TYPES = (
+    TS_SCALA_FUNCTION_DEFINITION,
+    TS_SCALA_FUNCTION_DECLARATION,
+)
+
+# (H) FQN node type tuples for PHP
+FQN_PHP_SCOPE_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_PHP_TRAIT_DECLARATION,
+    TS_PHP_NAMESPACE_DEFINITION,
+    TS_PROGRAM,
+)
+FQN_PHP_FUNCTION_TYPES = (
+    TS_PHP_FUNCTION_DEFINITION,
+    TS_PHP_METHOD_DECLARATION,
+    TS_PHP_ANONYMOUS_FUNCTION,
+    TS_PHP_ARROW_FUNCTION,
+)
+
+# (H) LANGUAGE_SPECS node type tuples for Python
+SPEC_PY_FUNCTION_TYPES = (TS_PY_FUNCTION_DEFINITION,)
+SPEC_PY_CLASS_TYPES = (TS_PY_CLASS_DEFINITION,)
+SPEC_PY_MODULE_TYPES = (TS_PY_MODULE,)
+SPEC_PY_CALL_TYPES = (TS_PY_CALL, TS_PY_WITH_STATEMENT)
+SPEC_PY_IMPORT_TYPES = (TS_PY_IMPORT_STATEMENT,)
+SPEC_PY_IMPORT_FROM_TYPES = (TS_PY_IMPORT_FROM_STATEMENT,)
+SPEC_PY_PACKAGE_INDICATORS = (PKG_INIT_PY,)
+
+# (H) LANGUAGE_SPECS node type tuples for JS/TS
+SPEC_JS_MODULE_TYPES = (TS_PROGRAM,)
+SPEC_JS_CALL_TYPES = (TS_CALL_EXPRESSION, TS_NEW_EXPRESSION)
+
+# (H) Derived node types for _js_get_name
+JS_NAME_NODE_TYPES = (
+    TS_FUNCTION_DECLARATION,
+    TS_CLASS_DECLARATION,
+    TS_METHOD_DEFINITION,
+    # (H) TS `namespace`/`module` block; its `name` field scopes nested classes.
+    TS_INTERNAL_MODULE,
+)
+
+# (H) Derived node types for _rust_get_name
+RS_TYPE_NODE_TYPES = (
+    TS_RS_STRUCT_ITEM,
+    TS_RS_ENUM_ITEM,
+    TS_RS_TRAIT_ITEM,
+    TS_RS_TYPE_ITEM,
+)
+RS_IDENT_NODE_TYPES = (TS_RS_FUNCTION_ITEM, TS_RS_MOD_ITEM)
+
+# (H) Derived node types for _cpp_get_name
+CPP_NAME_NODE_TYPES = (
+    CppNodeType.CLASS_SPECIFIER,
+    TS_STRUCT_SPECIFIER,
+    TS_ENUM_SPECIFIER,
+)
+
+# (H) Derived node types for _c_get_name
+C_NAME_NODE_TYPES = (
+    TS_STRUCT_SPECIFIER,
+    TS_UNION_SPECIFIER,
+    TS_ENUM_SPECIFIER,
+)
+
+# (H) LANGUAGE_SPECS node type tuples for Rust
+SPEC_RS_FUNCTION_TYPES = (
+    TS_RS_FUNCTION_ITEM,
+    TS_RS_FUNCTION_SIGNATURE_ITEM,
+    TS_RS_CLOSURE_EXPRESSION,
+    # (H) macro_rules! definitions register as Function nodes (the
+    # (H) cross-language decision: C/C++/Rust macros all map onto Function), so
+    # (H) macro_invocation call sites (already in SPEC_RS_CALL_TYPES) have a
+    # (H) first-party definition to resolve to.
+    TS_RS_MACRO_DEFINITION,
+)
+SPEC_RS_CLASS_TYPES = (
+    TS_RS_STRUCT_ITEM,
+    TS_RS_ENUM_ITEM,
+    TS_RS_UNION_ITEM,
+    TS_RS_TRAIT_ITEM,
+    TS_RS_IMPL_ITEM,
+    TS_RS_TYPE_ITEM,
+)
+SPEC_RS_MODULE_TYPES = (TS_RS_SOURCE_FILE, TS_RS_MOD_ITEM)
+SPEC_RS_CALL_TYPES = (TS_RS_CALL_EXPRESSION, TS_RS_MACRO_INVOCATION)
+SPEC_RS_IMPORT_TYPES = (TS_RS_USE_DECLARATION, TS_RS_EXTERN_CRATE_DECLARATION)
+SPEC_RS_IMPORT_FROM_TYPES = (TS_RS_USE_DECLARATION,)
+SPEC_RS_PACKAGE_INDICATORS = (PKG_CARGO_TOML,)
+
+# (H) LANGUAGE_SPECS node type tuples for Go
+SPEC_GO_FUNCTION_TYPES = (TS_GO_FUNCTION_DECLARATION, TS_GO_METHOD_DECLARATION)
+SPEC_GO_CLASS_TYPES = (TS_GO_TYPE_SPEC, TS_GO_TYPE_ALIAS)
+SPEC_GO_MODULE_TYPES = (TS_GO_SOURCE_FILE,)
+SPEC_GO_CALL_TYPES = (TS_GO_CALL_EXPRESSION,)
+SPEC_GO_IMPORT_TYPES = (TS_GO_IMPORT_DECLARATION,)
+
+# (H) LANGUAGE_SPECS node type tuples for Scala
+SPEC_SCALA_FUNCTION_TYPES = (
+    TS_SCALA_FUNCTION_DEFINITION,
+    TS_SCALA_FUNCTION_DECLARATION,
+)
+SPEC_SCALA_CLASS_TYPES = (
+    TS_SCALA_CLASS_DEFINITION,
+    TS_SCALA_OBJECT_DEFINITION,
+    TS_SCALA_TRAIT_DEFINITION,
+)
+SPEC_SCALA_MODULE_TYPES = (TS_SCALA_COMPILATION_UNIT,)
+SPEC_SCALA_CALL_TYPES = (
+    TS_SCALA_CALL_EXPRESSION,
+    TS_SCALA_GENERIC_FUNCTION,
+    TS_SCALA_FIELD_EXPRESSION,
+    TS_SCALA_INFIX_EXPRESSION,
+)
+SPEC_SCALA_IMPORT_TYPES = (TS_SCALA_IMPORT_DECLARATION,)
+
+# (H) LANGUAGE_SPECS node type tuples for Java
+SPEC_JAVA_FUNCTION_TYPES = (TS_METHOD_DECLARATION, TS_CONSTRUCTOR_DECLARATION)
+SPEC_JAVA_CLASS_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_ENUM_DECLARATION,
+    TS_JAVA_ANNOTATION_TYPE_DECLARATION,
+    TS_RECORD_DECLARATION,
+)
+SPEC_JAVA_MODULE_TYPES = (TS_PROGRAM,)
+SPEC_JAVA_CALL_TYPES = (TS_JAVA_METHOD_INVOCATION, TS_OBJECT_CREATION_EXPRESSION)
+SPEC_JAVA_IMPORT_TYPES = (TS_IMPORT_DECLARATION,)
+
+# (H) LANGUAGE_SPECS node type tuples for C++
+SPEC_CPP_FUNCTION_TYPES = (
+    TS_CPP_FUNCTION_DEFINITION,
+    TS_CPP_DECLARATION,
+    TS_CPP_FIELD_DECLARATION,
+    TS_CPP_TEMPLATE_DECLARATION,
+    TS_CPP_LAMBDA_EXPRESSION,
+)
+SPEC_CPP_CLASS_TYPES = (
+    CppNodeType.CLASS_SPECIFIER,
+    TS_STRUCT_SPECIFIER,
+    TS_UNION_SPECIFIER,
+    TS_ENUM_SPECIFIER,
+)
+SPEC_CPP_MODULE_TYPES = (
+    TS_CPP_TRANSLATION_UNIT,
+    TS_NAMESPACE_DEFINITION,
+    TS_CPP_LINKAGE_SPECIFICATION,
+    TS_CPP_DECLARATION,
+)
+SPEC_CPP_CALL_TYPES = (
+    TS_CPP_CALL_EXPRESSION,
+    TS_CPP_FIELD_EXPRESSION,
+    TS_CPP_SUBSCRIPT_EXPRESSION,
+    TS_CPP_NEW_EXPRESSION,
+    TS_CPP_DELETE_EXPRESSION,
+    TS_CPP_BINARY_EXPRESSION,
+    TS_CPP_UNARY_EXPRESSION,
+    TS_CPP_UPDATE_EXPRESSION,
+)
+SPEC_CPP_PACKAGE_INDICATORS = (
+    PKG_CMAKE_LISTS,
+    PKG_MAKEFILE,
+    PKG_VCXPROJ_GLOB,
+    PKG_CONANFILE,
+)
+
+# (H) FQN node type tuples for C
+FQN_C_SCOPE_TYPES = (
+    TS_CPP_TRANSLATION_UNIT,
+    TS_STRUCT_SPECIFIER,
+    TS_UNION_SPECIFIER,
+    TS_ENUM_SPECIFIER,
+)
+FQN_C_FUNCTION_TYPES = (TS_CPP_FUNCTION_DEFINITION,)
+
+# (H) LANGUAGE_SPECS node type tuples for C
+SPEC_C_FUNCTION_TYPES = (TS_CPP_FUNCTION_DEFINITION,)
+SPEC_C_CLASS_TYPES = (
+    TS_STRUCT_SPECIFIER,
+    TS_UNION_SPECIFIER,
+    TS_ENUM_SPECIFIER,
+)
+SPEC_C_MODULE_TYPES = (TS_CPP_TRANSLATION_UNIT,)
+SPEC_C_CALL_TYPES = (TS_CPP_CALL_EXPRESSION,)
+SPEC_C_PACKAGE_INDICATORS = (PKG_CMAKE_LISTS, PKG_MAKEFILE)
+
+# (H) LANGUAGE_SPECS node type tuples for PHP
+SPEC_PHP_FUNCTION_TYPES = (
+    TS_PHP_FUNCTION_DEFINITION,
+    TS_PHP_METHOD_DECLARATION,
+    TS_PHP_ANONYMOUS_FUNCTION,
+    TS_PHP_ARROW_FUNCTION,
+)
+SPEC_PHP_CLASS_TYPES = (
+    TS_CLASS_DECLARATION,
+    TS_INTERFACE_DECLARATION,
+    TS_PHP_TRAIT_DECLARATION,
+    TS_ENUM_DECLARATION,
+)
+SPEC_PHP_MODULE_TYPES = (TS_PROGRAM,)
+SPEC_PHP_CALL_TYPES = (
+    TS_PHP_FUNCTION_CALL_EXPRESSION,
+    TS_PHP_MEMBER_CALL_EXPRESSION,
+    TS_PHP_SCOPED_CALL_EXPRESSION,
+    TS_PHP_NULLSAFE_MEMBER_CALL_EXPRESSION,
+    TS_PHP_OBJECT_CREATION_EXPRESSION,
+)
+SPEC_PHP_IMPORT_TYPES = (TS_PHP_NAMESPACE_USE_DECLARATION,)
+SPEC_PHP_IMPORT_FROM_TYPES = (
+    TS_PHP_INCLUDE_EXPRESSION,
+    TS_PHP_INCLUDE_ONCE_EXPRESSION,
+    TS_PHP_REQUIRE_EXPRESSION,
+    TS_PHP_REQUIRE_ONCE_EXPRESSION,
+)
+
+# (H) LANGUAGE_SPECS node type tuples for Lua
+SPEC_LUA_FUNCTION_TYPES = (TS_LUA_FUNCTION_DECLARATION, TS_LUA_FUNCTION_DEFINITION)
+SPEC_LUA_CLASS_TYPES: tuple[str, ...] = ()
+SPEC_LUA_MODULE_TYPES = (TS_LUA_CHUNK,)
+SPEC_LUA_CALL_TYPES = (TS_LUA_FUNCTION_CALL,)
+SPEC_LUA_IMPORT_TYPES = (TS_LUA_FUNCTION_CALL,)
